refactor(web): split seedFoodsUk.ts by category
