+++ apps/web/src/modules/nutrition/lib/foodDb/seedFoodsUk.ts
@@ -1,1614 +1,45 @@
-import type { Macros } from "../macros";
+import type { SeedFood } from "./seeds/types";
 
-export interface SeedFood {
-  name: string;
-  per100: Macros;
-}
+export type { SeedFood };
+
+import { MEAT_AND_POULTRY } from "./seeds/meatAndPoultry";
+import { FISH_AND_SEAFOOD } from "./seeds/fishAndSeafood";
+import { EGGS } from "./seeds/eggs";
+import { DAIRY } from "./seeds/dairy";
+import { GRAINS } from "./seeds/grains";
+import { BAKERY } from "./seeds/bakery";
+import { LEGUMES } from "./seeds/legumes";
+import { VEGETABLES } from "./seeds/vegetables";
+import { FRUITS } from "./seeds/fruits";
+import { NUTS_AND_SEEDS } from "./seeds/nutsAndSeeds";
+import { OILS_AND_FATS } from "./seeds/oilsAndFats";
+import { SAUCES_AND_SPICES } from "./seeds/saucesAndSpices";
+import { SWEETS } from "./seeds/sweets";
+import { BEVERAGES } from "./seeds/beverages";
+import { SPORTS_NUTRITION } from "./seeds/sportsNutrition";
+import { READY_MEALS } from "./seeds/readyMeals";
+import { UKRAINIAN_CUISINE } from "./seeds/ukrainianCuisine";
+import { SALADS } from "./seeds/salads";
+import { FROZEN_SNACKS } from "./seeds/frozenSnacks";
 
 export const SEED_FOODS_UK: SeedFood[] = [
-  // М'ясо та птиця
-  {
-    name: "Куряча грудка",
-    per100: { kcal: 165, protein_g: 31, fat_g: 3.6, carbs_g: 0 },
-  },
-  {
-    name: "Куряче стегно",
-    per100: { kcal: 215, protein_g: 18, fat_g: 15, carbs_g: 0 },
-  },
-  {
-    name: "Куряче крило",
-    per100: { kcal: 222, protein_g: 18, fat_g: 15, carbs_g: 0 },
-  },
-  {
-    name: "Індичка (філе)",
-    per100: { kcal: 135, protein_g: 29, fat_g: 1.6, carbs_g: 0 },
-  },
-  {
-    name: "Яловичина",
-    per100: { kcal: 187, protein_g: 20, fat_g: 12, carbs_g: 0 },
-  },
-  {
-    name: "Яловичий фарш",
-    per100: { kcal: 254, protein_g: 17, fat_g: 20, carbs_g: 0 },
-  },
-  {
-    name: "Яловичий стейк",
-    per100: { kcal: 271, protein_g: 26, fat_g: 18, carbs_g: 0 },
-  },
-  {
-    name: "Свинина",
-    per100: { kcal: 242, protein_g: 17, fat_g: 19, carbs_g: 0 },
-  },
-  {
-    name: "Свинячий фарш",
-    per100: { kcal: 263, protein_g: 16, fat_g: 22, carbs_g: 0 },
-  },
-  {
-    name: "Курячий фарш",
-    per100: { kcal: 143, protein_g: 17, fat_g: 8, carbs_g: 0 },
-  },
-  {
-    name: "Шинка",
-    per100: { kcal: 163, protein_g: 17, fat_g: 10, carbs_g: 1.5 },
-  },
-  {
-    name: "Бекон",
-    per100: { kcal: 541, protein_g: 37, fat_g: 42, carbs_g: 1.5 },
-  },
-  {
-    name: "Ковбаса варена",
-    per100: { kcal: 260, protein_g: 12, fat_g: 23, carbs_g: 1 },
-  },
-  {
-    name: "Ковбаса копчена",
-    per100: { kcal: 420, protein_g: 17, fat_g: 38, carbs_g: 2 },
-  },
-  {
-    name: "Сосиски",
-    per100: { kcal: 256, protein_g: 10, fat_g: 23, carbs_g: 2 },
-  },
-  {
-    name: "Котлета м'ясна",
-    per100: { kcal: 230, protein_g: 15, fat_g: 18, carbs_g: 5 },
-  },
-  {
-    name: "Котлета куряча",
-    per100: { kcal: 200, protein_g: 17, fat_g: 14, carbs_g: 5 },
-  },
-  {
-    name: "Кролик",
-    per100: { kcal: 183, protein_g: 21, fat_g: 11, carbs_g: 0 },
-  },
-  {
-    name: "Качка",
-    per100: { kcal: 337, protein_g: 19, fat_g: 28, carbs_g: 0 },
-  },
-  {
-    name: "Шашлик свинячий",
-    per100: { kcal: 280, protein_g: 18, fat_g: 22, carbs_g: 2 },
-  },
-
-  // Риба та морепродукти
-  {
-    name: "Лосось",
-    per100: { kcal: 208, protein_g: 20, fat_g: 13, carbs_g: 0 },
-  },
-  {
-    name: "Тунець (у власному соку)",
-    per100: { kcal: 116, protein_g: 26, fat_g: 1, carbs_g: 0 },
-  },
-  {
-    name: "Оселедець",
-    per100: { kcal: 218, protein_g: 18, fat_g: 16, carbs_g: 0 },
-  },
-  {
-    name: "Скумбрія",
-    per100: { kcal: 205, protein_g: 19, fat_g: 14, carbs_g: 0 },
-  },
-  {
-    name: "Тріска",
-    per100: { kcal: 82, protein_g: 18, fat_g: 0.7, carbs_g: 0 },
-  },
-  {
-    name: "Форель",
-    per100: { kcal: 148, protein_g: 21, fat_g: 7, carbs_g: 0 },
-  },
-  {
-    name: "Горбуша",
-    per100: { kcal: 142, protein_g: 21, fat_g: 6, carbs_g: 0 },
-  },
-  { name: "Минтай", per100: { kcal: 82, protein_g: 18, fat_g: 1, carbs_g: 0 } },
-  { name: "Короп", per100: { kcal: 127, protein_g: 18, fat_g: 6, carbs_g: 0 } },
-  {
-    name: "Тілапія",
-    per100: { kcal: 128, protein_g: 26, fat_g: 3, carbs_g: 0 },
-  },
-  {
-    name: "Кальмар",
-    per100: { kcal: 92, protein_g: 18, fat_g: 1.4, carbs_g: 1.9 },
-  },
-  {
-    name: "Креветки",
-    per100: { kcal: 99, protein_g: 19, fat_g: 1.8, carbs_g: 0.9 },
-  },
-  {
-    name: "Мідії",
-    per100: { kcal: 86, protein_g: 12, fat_g: 2.2, carbs_g: 3.7 },
-  },
-
-  // Яйця
-  {
-    name: "Яйце куряче",
-    per100: { kcal: 143, protein_g: 13, fat_g: 10, carbs_g: 1.1 },
-  },
-  {
-    name: "Яйце варене",
-    per100: { kcal: 155, protein_g: 13, fat_g: 11, carbs_g: 1.1 },
-  },
-  {
-    name: "Яєчня",
-    per100: { kcal: 196, protein_g: 14, fat_g: 15, carbs_g: 1 },
-  },
-  {
-    name: "Омлет",
-    per100: { kcal: 154, protein_g: 10, fat_g: 12, carbs_g: 3 },
-  },
-
-  // Молочні продукти
-  {
-    name: "Молоко 2.5%",
-    per100: { kcal: 52, protein_g: 3, fat_g: 2.5, carbs_g: 4.7 },
-  },
-  {
-    name: "Молоко 1%",
-    per100: { kcal: 42, protein_g: 3.4, fat_g: 1, carbs_g: 5 },
-  },
-  {
-    name: "Молоко 3.2%",
-    per100: { kcal: 60, protein_g: 3, fat_g: 3.2, carbs_g: 4.7 },
-  },
-  {
-    name: "Кефір 2.5%",
-    per100: { kcal: 53, protein_g: 3, fat_g: 2.5, carbs_g: 4 },
-  },
-  {
-    name: "Ряжанка",
-    per100: { kcal: 85, protein_g: 3, fat_g: 4.2, carbs_g: 9 },
-  },
-  {
-    name: "Йогурт натуральний 2%",
-    per100: { kcal: 73, protein_g: 4.3, fat_g: 2, carbs_g: 8 },
-  },
-  {
-    name: "Сир кисломолочний 5%",
-    per100: { kcal: 121, protein_g: 17, fat_g: 5, carbs_g: 3 },
-  },
-  {
-    name: "Сир кисломолочний 0%",
-    per100: { kcal: 71, protein_g: 18, fat_g: 0.6, carbs_g: 3 },
-  },
-  {
-    name: "Сир твердий",
-    per100: { kcal: 360, protein_g: 25, fat_g: 28, carbs_g: 1 },
-  },
-  {
-    name: "Сир моцарела",
-    per100: { kcal: 280, protein_g: 18, fat_g: 22, carbs_g: 2 },
-  },
-  {
-    name: "Сир фета",
-    per100: { kcal: 264, protein_g: 14, fat_g: 21, carbs_g: 4 },
-  },
-  {
-    name: "Сир пармезан",
-    per100: { kcal: 431, protein_g: 38, fat_g: 29, carbs_g: 4 },
-  },
-  {
-    name: "Сир рікотта",
-    per100: { kcal: 174, protein_g: 11, fat_g: 13, carbs_g: 3 },
-  },
-  {
-    name: "Сметана 20%",
-    per100: { kcal: 204, protein_g: 2.8, fat_g: 20, carbs_g: 3.2 },
-  },
-  {
-    name: "Сметана 10%",
-    per100: { kcal: 115, protein_g: 3, fat_g: 10, carbs_g: 4 },
-  },
-  {
-    name: "Вершки 20%",
-    per100: { kcal: 207, protein_g: 2.5, fat_g: 20, carbs_g: 3.7 },
-  },
-  {
-    name: "Вершки 10%",
-    per100: { kcal: 119, protein_g: 2.7, fat_g: 10, carbs_g: 4.5 },
-  },
-  {
-    name: "Вершкове масло",
-    per100: { kcal: 717, protein_g: 0.9, fat_g: 81, carbs_g: 0.1 },
-  },
-  {
-    name: "Морозиво вершкове",
-    per100: { kcal: 207, protein_g: 3.5, fat_g: 11, carbs_g: 24 },
-  },
-
-  // Круп'яні та зернові
-  {
-    name: "Вівсянка (суха)",
-    per100: { kcal: 389, protein_g: 17, fat_g: 7, carbs_g: 66 },
-  },
-  {
-    name: "Вівсянка варена",
-    per100: { kcal: 71, protein_g: 2.5, fat_g: 1.4, carbs_g: 12 },
-  },
-  {
-    name: "Рис (сухий)",
-    per100: { kcal: 365, protein_g: 7, fat_g: 0.7, carbs_g: 80 },
-  },
-  {
-    name: "Рис варений",
-    per100: { kcal: 130, protein_g: 2.4, fat_g: 0.3, carbs_g: 28 },
-  },
-  {
-    name: "Гречка (суха)",
-    per100: { kcal: 343, protein_g: 13, fat_g: 3.4, carbs_g: 71 },
-  },
-  {
-    name: "Гречка варена",
-    per100: { kcal: 110, protein_g: 3.6, fat_g: 1.2, carbs_g: 20 },
-  },
-  {
-    name: "Макарони (сухі)",
-    per100: { kcal: 371, protein_g: 13, fat_g: 1.5, carbs_g: 75 },
-  },
-  {
-    name: "Паста варена",
-    per100: { kcal: 158, protein_g: 5.8, fat_g: 0.9, carbs_g: 31 },
-  },
-  {
-    name: "Булгур (сухий)",
-    per100: { kcal: 342, protein_g: 12, fat_g: 1.3, carbs_g: 76 },
-  },
-  {
-    name: "Пшоно (сухе)",
-    per100: { kcal: 378, protein_g: 11, fat_g: 3.9, carbs_g: 73 },
-  },
-  {
-    name: "Перлова крупа (суха)",
-    per100: { kcal: 352, protein_g: 9.9, fat_g: 1.2, carbs_g: 73 },
-  },
-  {
-    name: "Кіноа (суха)",
-    per100: { kcal: 368, protein_g: 14, fat_g: 6.1, carbs_g: 64 },
-  },
-  {
-    name: "Кус-кус (сухий)",
-    per100: { kcal: 376, protein_g: 13, fat_g: 0.6, carbs_g: 77 },
-  },
-  {
-    name: "Манна крупа",
-    per100: { kcal: 360, protein_g: 10, fat_g: 1, carbs_g: 73 },
-  },
-  {
-    name: "Кукурудзяна крупа",
-    per100: { kcal: 337, protein_g: 8, fat_g: 1.2, carbs_g: 75 },
-  },
-  {
-    name: "Вівсяні пластівці швидкі",
-    per100: { kcal: 352, protein_g: 11, fat_g: 6, carbs_g: 66 },
-  },
-  { name: "Мюслі", per100: { kcal: 367, protein_g: 9, fat_g: 6, carbs_g: 70 } },
-  {
-    name: "Гранола",
-    per100: { kcal: 471, protein_g: 9, fat_g: 16, carbs_g: 75 },
-  },
-  {
-    name: "Кукурудзяні пластівці",
-    per100: { kcal: 370, protein_g: 7, fat_g: 1, carbs_g: 84 },
-  },
-
-  // Хліб та випічка
-  {
-    name: "Хліб пшеничний",
-    per100: { kcal: 265, protein_g: 9, fat_g: 3.2, carbs_g: 49 },
-  },
-  {
-    name: "Хліб житній",
-    per100: { kcal: 259, protein_g: 9, fat_g: 3.3, carbs_g: 48 },
-  },
-  {
-    name: "Хліб цільнозерновий",
-    per100: { kcal: 247, protein_g: 9, fat_g: 3, carbs_g: 45 },
-  },
-  {
-    name: "Батон",
-    per100: { kcal: 262, protein_g: 8, fat_g: 3.2, carbs_g: 50 },
-  },
-  {
-    name: "Лаваш",
-    per100: { kcal: 277, protein_g: 9, fat_g: 1.2, carbs_g: 58 },
-  },
-  {
-    name: "Сухарі",
-    per100: { kcal: 347, protein_g: 11, fat_g: 4, carbs_g: 70 },
-  },
-  {
-    name: "Хлібці рисові",
-    per100: { kcal: 390, protein_g: 8, fat_g: 3, carbs_g: 84 },
-  },
-
-  // Бобові
-  {
-    name: "Квасоля (варена)",
-    per100: { kcal: 127, protein_g: 9, fat_g: 0.5, carbs_g: 23 },
-  },
-  {
-    name: "Сочевиця (варена)",
-    per100: { kcal: 116, protein_g: 9, fat_g: 0.4, carbs_g: 20 },
-  },
-  {
-    name: "Нут (варений)",
-    per100: { kcal: 164, protein_g: 9, fat_g: 2.6, carbs_g: 27 },
-  },
-  {
-    name: "Тофу",
-    per100: { kcal: 76, protein_g: 8, fat_g: 4.8, carbs_g: 1.9 },
-  },
-  {
-    name: "Едамаме",
-    per100: { kcal: 121, protein_g: 11, fat_g: 5.2, carbs_g: 8.9 },
-  },
-  {
-    name: "Горошок зелений",
-    per100: { kcal: 81, protein_g: 5, fat_g: 0.4, carbs_g: 14 },
-  },
-
-  // Овочі
-  {
-    name: "Картопля",
-    per100: { kcal: 77, protein_g: 2, fat_g: 0.1, carbs_g: 17 },
-  },
-  {
-    name: "Батат",
-    per100: { kcal: 86, protein_g: 1.6, fat_g: 0.1, carbs_g: 20 },
-  },
-  {
-    name: "Томат",
-    per100: { kcal: 18, protein_g: 0.9, fat_g: 0.2, carbs_g: 3.9 },
-  },
-  {
-    name: "Огірок",
-    per100: { kcal: 15, protein_g: 0.7, fat_g: 0.1, carbs_g: 3.6 },
-  },
-  {
-    name: "Перець солодкий",
-    per100: { kcal: 31, protein_g: 1, fat_g: 0.3, carbs_g: 6 },
-  },
-  {
-    name: "Морква",
-    per100: { kcal: 41, protein_g: 0.9, fat_g: 0.2, carbs_g: 10 },
-  },
-  {
-    name: "Цибуля",
-    per100: { kcal: 40, protein_g: 1.1, fat_g: 0.1, carbs_g: 9.3 },
-  },
-  {
-    name: "Цибуля зелена",
-    per100: { kcal: 32, protein_g: 1.8, fat_g: 0.2, carbs_g: 7 },
-  },
-  {
-    name: "Часник",
-    per100: { kcal: 149, protein_g: 6.4, fat_g: 0.5, carbs_g: 33 },
-  },
-  {
-    name: "Капуста білокачанна",
-    per100: { kcal: 25, protein_g: 1.3, fat_g: 0.1, carbs_g: 5.8 },
-  },
-  {
-    name: "Капуста пекінська",
-    per100: { kcal: 16, protein_g: 1.2, fat_g: 0.2, carbs_g: 2.5 },
-  },
-  {
-    name: "Цвітна капуста",
-    per100: { kcal: 25, protein_g: 1.9, fat_g: 0.3, carbs_g: 5 },
-  },
-  {
-    name: "Броколі",
-    per100: { kcal: 34, protein_g: 2.8, fat_g: 0.4, carbs_g: 7 },
-  },
-  {
-    name: "Шпинат",
-    per100: { kcal: 23, protein_g: 2.9, fat_g: 0.4, carbs_g: 3.6 },
-  },
-  {
-    name: "Салат листовий",
-    per100: { kcal: 15, protein_g: 1.4, fat_g: 0.2, carbs_g: 2.9 },
-  },
-  {
-    name: "Руккола",
-    per100: { kcal: 25, protein_g: 2.6, fat_g: 0.7, carbs_g: 3.7 },
-  },
-  {
-    name: "Буряк",
-    per100: { kcal: 43, protein_g: 1.6, fat_g: 0.2, carbs_g: 10 },
-  },
-  {
-    name: "Кабачок",
-    per100: { kcal: 17, protein_g: 1.2, fat_g: 0.3, carbs_g: 3.1 },
-  },
-  {
-    name: "Баклажан",
-    per100: { kcal: 25, protein_g: 1, fat_g: 0.2, carbs_g: 5.7 },
-  },
-  {
-    name: "Гарбуз",
-    per100: { kcal: 26, protein_g: 1, fat_g: 0.1, carbs_g: 6.5 },
-  },
-  {
-    name: "Кукурудза",
-    per100: { kcal: 86, protein_g: 3.3, fat_g: 1.4, carbs_g: 19 },
-  },
-  {
-    name: "Гриби (шампіньони)",
-    per100: { kcal: 22, protein_g: 3.1, fat_g: 0.3, carbs_g: 3.3 },
-  },
-  {
-    name: "Редиска",
-    per100: { kcal: 16, protein_g: 0.7, fat_g: 0.1, carbs_g: 3.4 },
-  },
-  {
-    name: "Петрушка",
-    per100: { kcal: 36, protein_g: 3, fat_g: 0.8, carbs_g: 6.3 },
-  },
-  {
-    name: "Кріп",
-    per100: { kcal: 43, protein_g: 3.5, fat_g: 1.1, carbs_g: 7 },
-  },
-  {
-    name: "Квашена капуста",
-    per100: { kcal: 19, protein_g: 0.9, fat_g: 0.1, carbs_g: 4 },
-  },
-  {
-    name: "Авокадо",
-    per100: { kcal: 160, protein_g: 2, fat_g: 15, carbs_g: 9 },
-  },
-  {
-    name: "Спаржа",
-    per100: { kcal: 20, protein_g: 2.2, fat_g: 0.1, carbs_g: 3.9 },
-  },
-
-  // Фрукти
-  {
-    name: "Яблуко",
-    per100: { kcal: 52, protein_g: 0.3, fat_g: 0.2, carbs_g: 14 },
-  },
-  {
-    name: "Груша",
-    per100: { kcal: 57, protein_g: 0.4, fat_g: 0.1, carbs_g: 15 },
-  },
-  {
-    name: "Банан",
-    per100: { kcal: 89, protein_g: 1.1, fat_g: 0.3, carbs_g: 23 },
-  },
-  {
-    name: "Апельсин",
-    per100: { kcal: 47, protein_g: 0.9, fat_g: 0.1, carbs_g: 12 },
-  },
-  {
-    name: "Мандарин",
-    per100: { kcal: 53, protein_g: 0.8, fat_g: 0.3, carbs_g: 13 },
-  },
-  {
-    name: "Лимон",
-    per100: { kcal: 29, protein_g: 1.1, fat_g: 0.3, carbs_g: 9 },
-  },
-  {
-    name: "Грейпфрут",
-    per100: { kcal: 42, protein_g: 0.8, fat_g: 0.1, carbs_g: 11 },
-  },
-  {
-    name: "Ківі",
-    per100: { kcal: 61, protein_g: 1.1, fat_g: 0.5, carbs_g: 14 },
-  },
-  {
-    name: "Манго",
-    per100: { kcal: 60, protein_g: 0.8, fat_g: 0.4, carbs_g: 15 },
-  },
-  {
-    name: "Ананас",
-    per100: { kcal: 50, protein_g: 0.5, fat_g: 0.1, carbs_g: 13 },
-  },
-  {
-    name: "Персик",
-    per100: { kcal: 39, protein_g: 0.9, fat_g: 0.3, carbs_g: 9.5 },
-  },
-  {
-    name: "Абрикос",
-    per100: { kcal: 48, protein_g: 1.4, fat_g: 0.4, carbs_g: 11 },
-  },
-  {
-    name: "Слива",
-    per100: { kcal: 46, protein_g: 0.7, fat_g: 0.3, carbs_g: 11 },
-  },
-  {
-    name: "Вишня",
-    per100: { kcal: 63, protein_g: 1, fat_g: 0.3, carbs_g: 16 },
-  },
-  {
-    name: "Черешня",
-    per100: { kcal: 63, protein_g: 1.1, fat_g: 0.4, carbs_g: 16 },
-  },
-  {
-    name: "Виноград",
-    per100: { kcal: 69, protein_g: 0.7, fat_g: 0.2, carbs_g: 18 },
-  },
-  {
-    name: "Полуниця",
-    per100: { kcal: 32, protein_g: 0.7, fat_g: 0.3, carbs_g: 7.7 },
-  },
-  {
-    name: "Малина",
-    per100: { kcal: 52, protein_g: 1.2, fat_g: 0.7, carbs_g: 12 },
-  },
-  {
-    name: "Чорниця",
-    per100: { kcal: 57, protein_g: 0.7, fat_g: 0.3, carbs_g: 14 },
-  },
-  {
-    name: "Смородина чорна",
-    per100: { kcal: 63, protein_g: 1.4, fat_g: 0.4, carbs_g: 15 },
-  },
-  {
-    name: "Смородина червона",
-    per100: { kcal: 56, protein_g: 1.4, fat_g: 0.2, carbs_g: 13 },
-  },
-  {
-    name: "Журавлина",
-    per100: { kcal: 46, protein_g: 0.4, fat_g: 0.1, carbs_g: 12 },
-  },
-  {
-    name: "Кавун",
-    per100: { kcal: 30, protein_g: 0.6, fat_g: 0.2, carbs_g: 7.5 },
-  },
-  {
-    name: "Диня",
-    per100: { kcal: 34, protein_g: 0.8, fat_g: 0.2, carbs_g: 8 },
-  },
-  {
-    name: "Хурма",
-    per100: { kcal: 70, protein_g: 0.6, fat_g: 0.4, carbs_g: 19 },
-  },
-  {
-    name: "Гранат",
-    per100: { kcal: 83, protein_g: 1.7, fat_g: 1.2, carbs_g: 19 },
-  },
-  {
-    name: "Інжир",
-    per100: { kcal: 74, protein_g: 0.8, fat_g: 0.3, carbs_g: 19 },
-  },
-  {
-    name: "Чорнослив",
-    per100: { kcal: 240, protein_g: 2.2, fat_g: 0.4, carbs_g: 64 },
-  },
-  {
-    name: "Родзинки",
-    per100: { kcal: 299, protein_g: 3.1, fat_g: 0.5, carbs_g: 79 },
-  },
-
-  // Горіхи та насіння
-  {
-    name: "Мигдаль",
-    per100: { kcal: 579, protein_g: 21, fat_g: 50, carbs_g: 22 },
-  },
-  {
-    name: "Горіхи волоські",
-    per100: { kcal: 654, protein_g: 15, fat_g: 65, carbs_g: 14 },
-  },
-  {
-    name: "Арахіс",
-    per100: { kcal: 567, protein_g: 26, fat_g: 49, carbs_g: 16 },
-  },
-  {
-    name: "Кешью",
-    per100: { kcal: 553, protein_g: 18, fat_g: 44, carbs_g: 30 },
-  },
-  {
-    name: "Фісташки",
-    per100: { kcal: 562, protein_g: 20, fat_g: 45, carbs_g: 28 },
-  },
-  {
-    name: "Фундук",
-    per100: { kcal: 628, protein_g: 15, fat_g: 61, carbs_g: 17 },
-  },
-  {
-    name: "Насіння соняшнику",
-    per100: { kcal: 584, protein_g: 21, fat_g: 51, carbs_g: 20 },
-  },
-  {
-    name: "Насіння гарбуза",
-    per100: { kcal: 559, protein_g: 30, fat_g: 49, carbs_g: 11 },
-  },
-  {
-    name: "Насіння льону",
-    per100: { kcal: 534, protein_g: 18, fat_g: 42, carbs_g: 29 },
-  },
-  {
-    name: "Кунжут",
-    per100: { kcal: 573, protein_g: 17, fat_g: 50, carbs_g: 23 },
-  },
-  {
-    name: "Паста арахісова",
-    per100: { kcal: 588, protein_g: 25, fat_g: 50, carbs_g: 20 },
-  },
-
-  // Олії та жири
-  {
-    name: "Оливкова олія",
-    per100: { kcal: 884, protein_g: 0, fat_g: 100, carbs_g: 0 },
-  },
-  {
-    name: "Олія соняшникова",
-    per100: { kcal: 884, protein_g: 0, fat_g: 100, carbs_g: 0 },
-  },
-  {
-    name: "Олія кокосова",
-    per100: { kcal: 862, protein_g: 0, fat_g: 100, carbs_g: 0 },
-  },
-
-  // Соуси та приправи
-  {
-    name: "Майонез",
-    per100: { kcal: 680, protein_g: 1, fat_g: 75, carbs_g: 2 },
-  },
-  {
-    name: "Кетчуп",
-    per100: { kcal: 100, protein_g: 1.5, fat_g: 0.1, carbs_g: 24 },
-  },
-  {
-    name: "Гірчиця",
-    per100: { kcal: 66, protein_g: 4, fat_g: 3.6, carbs_g: 6 },
-  },
-  {
-    name: "Соєвий соус",
-    per100: { kcal: 60, protein_g: 5.6, fat_g: 0, carbs_g: 8 },
-  },
-  {
-    name: "Томатна паста",
-    per100: { kcal: 82, protein_g: 3.9, fat_g: 0.5, carbs_g: 18 },
-  },
-  {
-    name: "Хумус",
-    per100: { kcal: 177, protein_g: 8, fat_g: 10, carbs_g: 16 },
-  },
-
-  // Цукор та солодощі
-  {
-    name: "Цукор",
-    per100: { kcal: 387, protein_g: 0, fat_g: 0, carbs_g: 100 },
-  },
-  { name: "Мед", per100: { kcal: 304, protein_g: 0.3, fat_g: 0, carbs_g: 82 } },
-  {
-    name: "Варення",
-    per100: { kcal: 272, protein_g: 0.3, fat_g: 0.1, carbs_g: 70 },
-  },
-  {
-    name: "Чорний шоколад",
-    per100: { kcal: 546, protein_g: 4.9, fat_g: 31, carbs_g: 61 },
-  },
-  {
-    name: "Шоколад молочний",
-    per100: { kcal: 535, protein_g: 6.9, fat_g: 29, carbs_g: 60 },
-  },
-  {
-    name: "Зефір",
-    per100: { kcal: 304, protein_g: 0.8, fat_g: 0, carbs_g: 79 },
-  },
-  {
-    name: "Мармелад",
-    per100: { kcal: 322, protein_g: 0.4, fat_g: 0.1, carbs_g: 82 },
-  },
-  {
-    name: "Печиво вівсяне",
-    per100: { kcal: 437, protein_g: 6.5, fat_g: 14, carbs_g: 71 },
-  },
-  {
-    name: "Вафлі",
-    per100: { kcal: 514, protein_g: 8, fat_g: 28, carbs_g: 62 },
-  },
-  {
-    name: "Чіпси",
-    per100: { kcal: 520, protein_g: 6, fat_g: 29, carbs_g: 60 },
-  },
-  {
-    name: "Попкорн",
-    per100: { kcal: 375, protein_g: 11, fat_g: 4, carbs_g: 78 },
-  },
-
-  // Напої
-  {
-    name: "Кава (без цукру)",
-    per100: { kcal: 1, protein_g: 0.1, fat_g: 0, carbs_g: 0 },
-  },
-  {
-    name: "Чай (без цукру)",
-    per100: { kcal: 1, protein_g: 0, fat_g: 0, carbs_g: 0 },
-  },
-  {
-    name: "Апельсиновий сік",
-    per100: { kcal: 45, protein_g: 0.7, fat_g: 0.2, carbs_g: 10 },
-  },
-  {
-    name: "Яблучний сік",
-    per100: { kcal: 46, protein_g: 0.1, fat_g: 0.1, carbs_g: 11 },
-  },
-
-  // Протеїни та спортивне харчування
-  {
-    name: "Сироватковий протеїн",
-    per100: { kcal: 370, protein_g: 80, fat_g: 4, carbs_g: 7 },
-  },
-  {
-    name: "Казеїн",
-    per100: { kcal: 360, protein_g: 78, fat_g: 3, carbs_g: 8 },
-  },
-  {
-    name: "Гейнер",
-    per100: { kcal: 380, protein_g: 25, fat_g: 4, carbs_g: 65 },
-  },
-
-  // Готові страви та напівфабрикати
-  {
-    name: "Вареники з картоплею",
-    per100: { kcal: 197, protein_g: 6, fat_g: 4, carbs_g: 34 },
-  },
-  {
-    name: "Вареники з сиром",
-    per100: { kcal: 222, protein_g: 9, fat_g: 5, carbs_g: 36 },
-  },
-  {
-    name: "Пельмені",
-    per100: { kcal: 235, protein_g: 12, fat_g: 10, carbs_g: 26 },
-  },
-  { name: "Борщ", per100: { kcal: 58, protein_g: 3, fat_g: 2.5, carbs_g: 7 } },
-  {
-    name: "Пюре картопляне",
-    per100: { kcal: 87, protein_g: 2, fat_g: 3, carbs_g: 14 },
-  },
-  { name: "Плов", per100: { kcal: 190, protein_g: 8, fat_g: 8, carbs_g: 24 } },
-  {
-    name: "Тушкована капуста",
-    per100: { kcal: 75, protein_g: 2, fat_g: 4, carbs_g: 9 },
-  },
-  {
-    name: "Піца",
-    per100: { kcal: 266, protein_g: 11, fat_g: 10, carbs_g: 33 },
-  },
-  {
-    name: "Суші/ролл",
-    per100: { kcal: 170, protein_g: 5, fat_g: 2.5, carbs_g: 32 },
-  },
-
-  // М'ясо та делікатеси (розширено)
-  {
-    name: "Сало",
-    per100: { kcal: 797, protein_g: 2.4, fat_g: 89, carbs_g: 0 },
-  },
-  {
-    name: "Хамон",
-    per100: { kcal: 241, protein_g: 31, fat_g: 13, carbs_g: 0 },
-  },
-  {
-    name: "Пармська шинка",
-    per100: { kcal: 250, protein_g: 26, fat_g: 16, carbs_g: 0.3 },
-  },
-  {
-    name: "Саламі",
-    per100: { kcal: 407, protein_g: 22, fat_g: 34, carbs_g: 1.6 },
-  },
-  {
-    name: "Пепероні",
-    per100: { kcal: 494, protein_g: 23, fat_g: 44, carbs_g: 1.2 },
-  },
-  {
-    name: "Бастурма",
-    per100: { kcal: 240, protein_g: 50, fat_g: 4, carbs_g: 0 },
-  },
-  {
-    name: "Холодець",
-    per100: { kcal: 120, protein_g: 12, fat_g: 8, carbs_g: 0.5 },
-  },
-  {
-    name: "Печінка куряча",
-    per100: { kcal: 137, protein_g: 17, fat_g: 6.5, carbs_g: 1.4 },
-  },
-  {
-    name: "Печінка яловича",
-    per100: { kcal: 135, protein_g: 20, fat_g: 3.6, carbs_g: 5.3 },
-  },
-  {
-    name: "Серце куряче",
-    per100: { kcal: 153, protein_g: 16, fat_g: 9.3, carbs_g: 0.7 },
-  },
-  {
-    name: "Шлунки курячі",
-    per100: { kcal: 94, protein_g: 18, fat_g: 2.1, carbs_g: 0 },
-  },
-  {
-    name: "Баранина",
-    per100: { kcal: 294, protein_g: 25, fat_g: 21, carbs_g: 0 },
-  },
-  {
-    name: "Гусятина",
-    per100: { kcal: 305, protein_g: 16, fat_g: 27, carbs_g: 0 },
-  },
-  {
-    name: "Телятина",
-    per100: { kcal: 172, protein_g: 20, fat_g: 10, carbs_g: 0 },
-  },
-  {
-    name: "Філе індички",
-    per100: { kcal: 112, protein_g: 24, fat_g: 1.7, carbs_g: 0 },
-  },
-  {
-    name: "Куряча ніжка",
-    per100: { kcal: 184, protein_g: 20, fat_g: 11, carbs_g: 0 },
-  },
-  {
-    name: "Котлета по-київськи",
-    per100: { kcal: 290, protein_g: 15, fat_g: 22, carbs_g: 9 },
-  },
-  {
-    name: "Нагетси курячі",
-    per100: { kcal: 297, protein_g: 15, fat_g: 19, carbs_g: 16 },
-  },
-
-  // Риба та морепродукти (розширено)
-  {
-    name: "Ікра червона",
-    per100: { kcal: 263, protein_g: 32, fat_g: 15, carbs_g: 0 },
-  },
-  {
-    name: "Ікра чорна",
-    per100: { kcal: 264, protein_g: 25, fat_g: 18, carbs_g: 0 },
-  },
-  {
-    name: "Крабові палички",
-    per100: { kcal: 88, protein_g: 7, fat_g: 1, carbs_g: 14 },
-  },
-  {
-    name: "Оселедець солений",
-    per100: { kcal: 217, protein_g: 19, fat_g: 15, carbs_g: 0 },
-  },
-  {
-    name: "Шпроти в олії",
-    per100: { kcal: 363, protein_g: 17, fat_g: 32, carbs_g: 0.4 },
-  },
-  {
-    name: "Сардина в олії",
-    per100: { kcal: 208, protein_g: 25, fat_g: 11, carbs_g: 0 },
-  },
-  {
-    name: "Судак",
-    per100: { kcal: 84, protein_g: 19, fat_g: 1.1, carbs_g: 0 },
-  },
-  {
-    name: "Сом",
-    per100: { kcal: 115, protein_g: 16, fat_g: 5.3, carbs_g: 0 },
-  },
-  {
-    name: "Хек",
-    per100: { kcal: 86, protein_g: 16, fat_g: 2.2, carbs_g: 0 },
-  },
-  {
-    name: "Оселедець маринований",
-    per100: { kcal: 262, protein_g: 17, fat_g: 21, carbs_g: 1 },
-  },
-  {
-    name: "Устриці",
-    per100: { kcal: 68, protein_g: 9, fat_g: 2.3, carbs_g: 3.9 },
-  },
-  {
-    name: "Восьминіг",
-    per100: { kcal: 82, protein_g: 15, fat_g: 1, carbs_g: 2.2 },
-  },
-
-  // Молочні продукти (розширено)
-  {
-    name: "Молоко безлактозне",
-    per100: { kcal: 46, protein_g: 3.3, fat_g: 1.5, carbs_g: 4.7 },
-  },
-  {
-    name: "Молоко знежирене",
-    per100: { kcal: 35, protein_g: 3.4, fat_g: 0.1, carbs_g: 5 },
-  },
-  {
-    name: "Молоко кокосове",
-    per100: { kcal: 230, protein_g: 2.3, fat_g: 24, carbs_g: 5.5 },
-  },
-  {
-    name: "Молоко мигдалеве",
-    per100: { kcal: 17, protein_g: 0.6, fat_g: 1.1, carbs_g: 1.5 },
-  },
-  {
-    name: "Молоко вівсяне",
-    per100: { kcal: 45, protein_g: 0.8, fat_g: 1.5, carbs_g: 6.7 },
-  },
-  {
-    name: "Молоко соєве",
-    per100: { kcal: 45, protein_g: 3.3, fat_g: 1.8, carbs_g: 3.3 },
-  },
-  {
-    name: "Йогурт грецький 2%",
-    per100: { kcal: 73, protein_g: 10, fat_g: 2, carbs_g: 3.6 },
-  },
-  {
-    name: "Йогурт грецький 0%",
-    per100: { kcal: 59, protein_g: 10, fat_g: 0.4, carbs_g: 3.6 },
-  },
-  {
-    name: "Йогурт питний",
-    per100: { kcal: 72, protein_g: 2.8, fat_g: 1.5, carbs_g: 12 },
-  },
-  {
-    name: "Сир кисломолочний 9%",
-    per100: { kcal: 159, protein_g: 17, fat_g: 9, carbs_g: 2 },
-  },
-  {
-    name: "Сир кисломолочний 18%",
-    per100: { kcal: 236, protein_g: 14, fat_g: 18, carbs_g: 2.8 },
-  },
-  {
-    name: "Сирок глазурований",
-    per100: { kcal: 407, protein_g: 8.5, fat_g: 27, carbs_g: 32 },
-  },
-  {
-    name: "Сир чеддер",
-    per100: { kcal: 403, protein_g: 25, fat_g: 33, carbs_g: 1.3 },
-  },
-  {
-    name: "Сир гауда",
-    per100: { kcal: 356, protein_g: 25, fat_g: 27, carbs_g: 2.2 },
-  },
-  {
-    name: "Сир камамбер",
-    per100: { kcal: 299, protein_g: 20, fat_g: 24, carbs_g: 0.5 },
-  },
-  {
-    name: "Сир брі",
-    per100: { kcal: 334, protein_g: 21, fat_g: 28, carbs_g: 0.5 },
-  },
-  {
-    name: "Сир дор блю",
-    per100: { kcal: 353, protein_g: 21, fat_g: 29, carbs_g: 2.3 },
-  },
-  {
-    name: "Сир бринза",
-    per100: { kcal: 262, protein_g: 18, fat_g: 20, carbs_g: 2 },
-  },
-  {
-    name: "Сир сулугуні",
-    per100: { kcal: 290, protein_g: 20, fat_g: 22, carbs_g: 0.4 },
-  },
-  {
-    name: "Сир вершковий",
-    per100: { kcal: 342, protein_g: 6, fat_g: 34, carbs_g: 4 },
-  },
-  {
-    name: "Маскарпоне",
-    per100: { kcal: 430, protein_g: 5, fat_g: 45, carbs_g: 4.5 },
-  },
-  {
-    name: "Сметана 15%",
-    per100: { kcal: 158, protein_g: 2.8, fat_g: 15, carbs_g: 3.6 },
-  },
-  {
-    name: "Сметана 30%",
-    per100: { kcal: 293, protein_g: 2.4, fat_g: 30, carbs_g: 3.1 },
-  },
-  {
-    name: "Ацидофілін",
-    per100: { kcal: 57, protein_g: 2.8, fat_g: 3.2, carbs_g: 3.8 },
-  },
-  {
-    name: "Сироватка",
-    per100: { kcal: 20, protein_g: 0.8, fat_g: 0.2, carbs_g: 3.5 },
-  },
-
-  // Яйця (розширено)
-  {
-    name: "Яйце перепелине",
-    per100: { kcal: 158, protein_g: 13, fat_g: 11, carbs_g: 0.6 },
-  },
-  {
-    name: "Яєчний білок",
-    per100: { kcal: 52, protein_g: 11, fat_g: 0.2, carbs_g: 0.7 },
-  },
-  {
-    name: "Яєчний жовток",
-    per100: { kcal: 322, protein_g: 16, fat_g: 27, carbs_g: 3.6 },
-  },
-
-  // Крупи та варені гарніри
-  {
-    name: "Пшоно варене",
-    per100: { kcal: 90, protein_g: 3, fat_g: 0.7, carbs_g: 17 },
-  },
-  {
-    name: "Булгур варений",
-    per100: { kcal: 83, protein_g: 3.1, fat_g: 0.2, carbs_g: 19 },
-  },
-  {
-    name: "Кіноа варена",
-    per100: { kcal: 120, protein_g: 4.4, fat_g: 1.9, carbs_g: 21 },
-  },
-  {
-    name: "Кус-кус варений",
-    per100: { kcal: 112, protein_g: 3.8, fat_g: 0.2, carbs_g: 23 },
-  },
-  {
-    name: "Перлова варена",
-    per100: { kcal: 123, protein_g: 2.3, fat_g: 0.4, carbs_g: 28 },
-  },
-  {
-    name: "Рис бурий варений",
-    per100: { kcal: 112, protein_g: 2.6, fat_g: 0.9, carbs_g: 23 },
-  },
-  {
-    name: "Амарант",
-    per100: { kcal: 371, protein_g: 14, fat_g: 7, carbs_g: 65 },
-  },
-  {
-    name: "Полента",
-    per100: { kcal: 70, protein_g: 1.5, fat_g: 0.3, carbs_g: 15 },
-  },
-
-  // Хлібобулочні та випічка
-  {
-    name: "Багет",
-    per100: { kcal: 274, protein_g: 9, fat_g: 1.3, carbs_g: 56 },
-  },
-  {
-    name: "Чіабата",
-    per100: { kcal: 271, protein_g: 8.9, fat_g: 3.5, carbs_g: 51 },
-  },
-  {
-    name: "Тортилья пшенична",
-    per100: { kcal: 310, protein_g: 8, fat_g: 8, carbs_g: 51 },
-  },
-  {
-    name: "Хлібець гречаний",
-    per100: { kcal: 320, protein_g: 11, fat_g: 2.2, carbs_g: 63 },
-  },
-  {
-    name: "Булочка для бургера",
-    per100: { kcal: 279, protein_g: 9, fat_g: 5, carbs_g: 50 },
-  },
-  {
-    name: "Круасан",
-    per100: { kcal: 406, protein_g: 8.2, fat_g: 21, carbs_g: 46 },
-  },
-  {
-    name: "Булочка здобна",
-    per100: { kcal: 339, protein_g: 7.8, fat_g: 9.2, carbs_g: 56 },
-  },
-  {
-    name: "Пиріжок з капустою",
-    per100: { kcal: 223, protein_g: 5, fat_g: 6, carbs_g: 36 },
-  },
-  {
-    name: "Пиріжок з м'ясом",
-    per100: { kcal: 253, protein_g: 10, fat_g: 9, carbs_g: 32 },
-  },
-  {
-    name: "Маффін",
-    per100: { kcal: 377, protein_g: 5.5, fat_g: 15, carbs_g: 55 },
-  },
-  {
-    name: "Пончик",
-    per100: { kcal: 452, protein_g: 4.9, fat_g: 25, carbs_g: 51 },
-  },
-  {
-    name: "Еклер",
-    per100: { kcal: 262, protein_g: 6.3, fat_g: 16, carbs_g: 24 },
-  },
-
-  // Соуси, приправи, жири (розширено)
-  {
-    name: "Песто",
-    per100: { kcal: 431, protein_g: 5, fat_g: 44, carbs_g: 6 },
-  },
-  {
-    name: "Соус терияки",
-    per100: { kcal: 90, protein_g: 5.9, fat_g: 0, carbs_g: 15 },
-  },
-  {
-    name: "Соус барбекю",
-    per100: { kcal: 172, protein_g: 1, fat_g: 0.5, carbs_g: 41 },
-  },
-  {
-    name: "Соус солодкий чилі",
-    per100: { kcal: 229, protein_g: 0.8, fat_g: 0.3, carbs_g: 56 },
-  },
-  {
-    name: "Соус тартар",
-    per100: { kcal: 389, protein_g: 1.4, fat_g: 38, carbs_g: 9.6 },
-  },
-  {
-    name: "Тахіні",
-    per100: { kcal: 595, protein_g: 17, fat_g: 54, carbs_g: 21 },
-  },
-  {
-    name: "Аджика",
-    per100: { kcal: 59, protein_g: 1, fat_g: 3.7, carbs_g: 5.8 },
-  },
-  {
-    name: "Ткемалі",
-    per100: { kcal: 89, protein_g: 1, fat_g: 0.2, carbs_g: 20 },
-  },
-  {
-    name: "Маргарин",
-    per100: { kcal: 717, protein_g: 0.2, fat_g: 80, carbs_g: 0.9 },
-  },
-  {
-    name: "Топлене масло (гі)",
-    per100: { kcal: 900, protein_g: 0, fat_g: 100, carbs_g: 0 },
-  },
-
-  // Напої (розширено)
-  {
-    name: "Лате",
-    per100: { kcal: 56, protein_g: 3, fat_g: 3, carbs_g: 4.4 },
-  },
-  {
-    name: "Капучіно",
-    per100: { kcal: 37, protein_g: 2, fat_g: 2, carbs_g: 3 },
-  },
-  {
-    name: "Американо",
-    per100: { kcal: 2, protein_g: 0.1, fat_g: 0, carbs_g: 0 },
-  },
-  {
-    name: "Какао з молоком",
-    per100: { kcal: 76, protein_g: 3, fat_g: 3.5, carbs_g: 8 },
-  },
-  {
-    name: "Матча (лате)",
-    per100: { kcal: 55, protein_g: 2.5, fat_g: 2.5, carbs_g: 5.5 },
-  },
-  {
-    name: "Кола",
-    per100: { kcal: 42, protein_g: 0, fat_g: 0, carbs_g: 10.6 },
-  },
-  {
-    name: "Кола Zero",
-    per100: { kcal: 0.3, protein_g: 0, fat_g: 0, carbs_g: 0 },
-  },
-  {
-    name: "Спрайт",
-    per100: { kcal: 40, protein_g: 0, fat_g: 0, carbs_g: 10 },
-  },
-  {
-    name: "Фанта",
-    per100: { kcal: 48, protein_g: 0, fat_g: 0, carbs_g: 12 },
-  },
-  {
-    name: "Енергетик Red Bull",
-    per100: { kcal: 46, protein_g: 0.4, fat_g: 0, carbs_g: 11 },
-  },
-  {
-    name: "Квас",
-    per100: { kcal: 27, protein_g: 0.2, fat_g: 0, carbs_g: 5.2 },
-  },
-  {
-    name: "Узвар",
-    per100: { kcal: 48, protein_g: 0.3, fat_g: 0, carbs_g: 12 },
-  },
-  {
-    name: "Компот (без цукру)",
-    per100: { kcal: 24, protein_g: 0.1, fat_g: 0, carbs_g: 6 },
-  },
-  {
-    name: "Пиво світле 4.5%",
-    per100: { kcal: 43, protein_g: 0.5, fat_g: 0, carbs_g: 3.6 },
-  },
-  {
-    name: "Пиво темне",
-    per100: { kcal: 58, protein_g: 0.5, fat_g: 0, carbs_g: 4 },
-  },
-  {
-    name: "Вино сухе",
-    per100: { kcal: 82, protein_g: 0.1, fat_g: 0, carbs_g: 2.6 },
-  },
-  {
-    name: "Вино напівсолодке",
-    per100: { kcal: 98, protein_g: 0.2, fat_g: 0, carbs_g: 5.5 },
-  },
-  {
-    name: "Шампанське",
-    per100: { kcal: 88, protein_g: 0.2, fat_g: 0, carbs_g: 3 },
-  },
-  {
-    name: "Горілка 40%",
-    per100: { kcal: 231, protein_g: 0, fat_g: 0, carbs_g: 0.1 },
-  },
-  {
-    name: "Віскі 40%",
-    per100: { kcal: 250, protein_g: 0, fat_g: 0, carbs_g: 0 },
-  },
-
-  // Фрукти та сухофрукти (розширено)
-  {
-    name: "Кокос",
-    per100: { kcal: 354, protein_g: 3.3, fat_g: 33, carbs_g: 15 },
-  },
-  {
-    name: "Фініки",
-    per100: { kcal: 282, protein_g: 2.5, fat_g: 0.4, carbs_g: 75 },
-  },
-  {
-    name: "Курага",
-    per100: { kcal: 241, protein_g: 3.4, fat_g: 0.5, carbs_g: 63 },
-  },
-  {
-    name: "Фейхоа",
-    per100: { kcal: 49, protein_g: 0.7, fat_g: 0.6, carbs_g: 11 },
-  },
-  {
-    name: "Нектарин",
-    per100: { kcal: 44, protein_g: 1.1, fat_g: 0.3, carbs_g: 11 },
-  },
-  {
-    name: "Папайя",
-    per100: { kcal: 43, protein_g: 0.5, fat_g: 0.3, carbs_g: 11 },
-  },
-  {
-    name: "Агрус",
-    per100: { kcal: 44, protein_g: 0.9, fat_g: 0.6, carbs_g: 9.8 },
-  },
-  {
-    name: "Обліпиха",
-    per100: { kcal: 82, protein_g: 1.2, fat_g: 5.4, carbs_g: 5.7 },
-  },
-  {
-    name: "Лохина",
-    per100: { kcal: 57, protein_g: 0.7, fat_g: 0.3, carbs_g: 14 },
-  },
-  {
-    name: "Калина",
-    per100: { kcal: 26, protein_g: 0.4, fat_g: 0, carbs_g: 7 },
-  },
-
-  // Овочі (розширено)
-  {
-    name: "Імбир свіжий",
-    per100: { kcal: 80, protein_g: 1.8, fat_g: 0.8, carbs_g: 18 },
-  },
-  {
-    name: "Селера (черешковий)",
-    per100: { kcal: 16, protein_g: 0.7, fat_g: 0.2, carbs_g: 3 },
-  },
-  {
-    name: "Селера (корінь)",
-    per100: { kcal: 42, protein_g: 1.5, fat_g: 0.3, carbs_g: 9.2 },
-  },
-  {
-    name: "Ріпа",
-    per100: { kcal: 28, protein_g: 0.9, fat_g: 0.1, carbs_g: 6.4 },
-  },
-  {
-    name: "Редька",
-    per100: { kcal: 20, protein_g: 1.2, fat_g: 0.2, carbs_g: 3.4 },
-  },
-  {
-    name: "Перець чилі",
-    per100: { kcal: 40, protein_g: 1.9, fat_g: 0.4, carbs_g: 8.8 },
-  },
-  {
-    name: "Гриби білі",
-    per100: { kcal: 34, protein_g: 3.7, fat_g: 1.7, carbs_g: 1.1 },
-  },
-  {
-    name: "Гриби лисички",
-    per100: { kcal: 38, protein_g: 1.5, fat_g: 1.1, carbs_g: 6.9 },
-  },
-  {
-    name: "Гриби печериці мариновані",
-    per100: { kcal: 24, protein_g: 2.8, fat_g: 0.4, carbs_g: 0.1 },
-  },
-  {
-    name: "Помідор черрі",
-    per100: { kcal: 18, protein_g: 0.9, fat_g: 0.2, carbs_g: 3.9 },
-  },
-  {
-    name: "Томат в'ялений",
-    per100: { kcal: 258, protein_g: 14, fat_g: 3, carbs_g: 56 },
-  },
-  {
-    name: "Кукурудза консервована",
-    per100: { kcal: 58, protein_g: 2.2, fat_g: 0.6, carbs_g: 11 },
-  },
-  {
-    name: "Горошок консервований",
-    per100: { kcal: 53, protein_g: 3.1, fat_g: 0.2, carbs_g: 9.8 },
-  },
-  {
-    name: "Маслини чорні",
-    per100: { kcal: 115, protein_g: 0.8, fat_g: 11, carbs_g: 6.3 },
-  },
-  {
-    name: "Оливки зелені",
-    per100: { kcal: 145, protein_g: 1, fat_g: 15, carbs_g: 3.8 },
-  },
-  {
-    name: "Каперси",
-    per100: { kcal: 23, protein_g: 2.4, fat_g: 0.9, carbs_g: 5 },
-  },
-  {
-    name: "Корнішони",
-    per100: { kcal: 15, protein_g: 0.7, fat_g: 0.1, carbs_g: 2.4 },
-  },
-  {
-    name: "Огірок солоний",
-    per100: { kcal: 11, protein_g: 0.8, fat_g: 0.1, carbs_g: 1.7 },
-  },
-
-  // Фастфуд / готові страви (розширено)
-  {
-    name: "Бургер",
-    per100: { kcal: 295, protein_g: 17, fat_g: 14, carbs_g: 27 },
-  },
-  {
-    name: "Чізбургер",
-    per100: { kcal: 303, protein_g: 16, fat_g: 15, carbs_g: 26 },
-  },
-  {
-    name: "Хот-дог",
-    per100: { kcal: 247, protein_g: 10, fat_g: 15, carbs_g: 18 },
-  },
-  {
-    name: "Шаурма",
-    per100: { kcal: 215, protein_g: 12, fat_g: 11, carbs_g: 18 },
-  },
-  {
-    name: "Картопля фрі",
-    per100: { kcal: 312, protein_g: 3.4, fat_g: 15, carbs_g: 41 },
-  },
-  {
-    name: "Картопля по-селянськи",
-    per100: { kcal: 147, protein_g: 2.5, fat_g: 7, carbs_g: 19 },
-  },
-  {
-    name: "Сендвіч з куркою",
-    per100: { kcal: 221, protein_g: 13, fat_g: 9, carbs_g: 22 },
-  },
-  {
-    name: "Ролл з лососем",
-    per100: { kcal: 184, protein_g: 7, fat_g: 4, carbs_g: 30 },
-  },
-  {
-    name: "Піца Маргарита",
-    per100: { kcal: 251, protein_g: 11, fat_g: 8, carbs_g: 33 },
-  },
-  {
-    name: "Піца Пепероні",
-    per100: { kcal: 296, protein_g: 12, fat_g: 13, carbs_g: 32 },
-  },
-  {
-    name: "Піца 4 сири",
-    per100: { kcal: 313, protein_g: 14, fat_g: 15, carbs_g: 30 },
-  },
-  {
-    name: "Паста карбонара",
-    per100: { kcal: 200, protein_g: 9, fat_g: 10, carbs_g: 20 },
-  },
-  {
-    name: "Паста болоньєзе",
-    per100: { kcal: 160, protein_g: 8, fat_g: 5, carbs_g: 21 },
-  },
-  {
-    name: "Лазанья",
-    per100: { kcal: 135, protein_g: 8, fat_g: 5, carbs_g: 14 },
-  },
-  {
-    name: "Рагу овочеве",
-    per100: { kcal: 56, protein_g: 1.8, fat_g: 2, carbs_g: 8 },
-  },
-  {
-    name: "Тушкована яловичина",
-    per100: { kcal: 232, protein_g: 16, fat_g: 17, carbs_g: 3 },
-  },
-
-  // Українська кухня (розширено)
-  {
-    name: "Борщ зелений",
-    per100: { kcal: 49, protein_g: 2.5, fat_g: 2, carbs_g: 5 },
-  },
-  {
-    name: "Куряча локшина",
-    per100: { kcal: 45, protein_g: 3, fat_g: 1.5, carbs_g: 5 },
-  },
-  {
-    name: "Суп-харчо",
-    per100: { kcal: 86, protein_g: 6, fat_g: 4, carbs_g: 7 },
-  },
-  {
-    name: "Розсольник",
-    per100: { kcal: 49, protein_g: 2.5, fat_g: 2, carbs_g: 5.5 },
-  },
-  {
-    name: "Солянка м'ясна",
-    per100: { kcal: 105, protein_g: 6, fat_g: 7, carbs_g: 4 },
-  },
-  {
-    name: "Окрошка",
-    per100: { kcal: 53, protein_g: 2.7, fat_g: 2, carbs_g: 6 },
-  },
-  {
-    name: "Гороховий суп",
-    per100: { kcal: 66, protein_g: 4, fat_g: 2, carbs_g: 8 },
-  },
-  {
-    name: "Голубці",
-    per100: { kcal: 130, protein_g: 7, fat_g: 7, carbs_g: 10 },
-  },
-  {
-    name: "Деруни",
-    per100: { kcal: 170, protein_g: 3, fat_g: 10, carbs_g: 17 },
-  },
-  {
-    name: "Млинці",
-    per100: { kcal: 189, protein_g: 5, fat_g: 7, carbs_g: 27 },
-  },
-  {
-    name: "Налисники з сиром",
-    per100: { kcal: 219, protein_g: 9, fat_g: 10, carbs_g: 25 },
-  },
-  {
-    name: "Сирники",
-    per100: { kcal: 220, protein_g: 14, fat_g: 8, carbs_g: 23 },
-  },
-  {
-    name: "Галушки",
-    per100: { kcal: 165, protein_g: 5, fat_g: 3, carbs_g: 30 },
-  },
-  {
-    name: "Вареники з вишнею",
-    per100: { kcal: 165, protein_g: 4, fat_g: 1, carbs_g: 35 },
-  },
-  {
-    name: "Вареники з м'ясом",
-    per100: { kcal: 244, protein_g: 13, fat_g: 8, carbs_g: 30 },
-  },
-  {
-    name: "Котлета домашня",
-    per100: { kcal: 221, protein_g: 14, fat_g: 14, carbs_g: 10 },
-  },
-  {
-    name: "Кабачкова ікра",
-    per100: { kcal: 91, protein_g: 1.3, fat_g: 6.8, carbs_g: 6.4 },
-  },
-  {
-    name: "Баклажанна ікра",
-    per100: { kcal: 88, protein_g: 1.7, fat_g: 6.7, carbs_g: 6.1 },
-  },
-
-  // Салати (розширено)
-  {
-    name: "Олів'є",
-    per100: { kcal: 198, protein_g: 5, fat_g: 14, carbs_g: 12 },
-  },
-  {
-    name: "Вінегрет",
-    per100: { kcal: 76, protein_g: 1.5, fat_g: 4, carbs_g: 8 },
-  },
-  {
-    name: "Оселедець під шубою",
-    per100: { kcal: 193, protein_g: 5, fat_g: 15, carbs_g: 9 },
-  },
-  {
-    name: "Цезар з куркою",
-    per100: { kcal: 197, protein_g: 11, fat_g: 15, carbs_g: 5 },
-  },
-  {
-    name: "Грецький салат",
-    per100: { kcal: 99, protein_g: 3, fat_g: 8, carbs_g: 5 },
-  },
-  {
-    name: "Капрезе",
-    per100: { kcal: 212, protein_g: 11, fat_g: 17, carbs_g: 3 },
-  },
-  {
-    name: "Салат з тунцем",
-    per100: { kcal: 135, protein_g: 14, fat_g: 7, carbs_g: 4 },
-  },
-  {
-    name: "Салат із буряка",
-    per100: { kcal: 72, protein_g: 1.5, fat_g: 4, carbs_g: 8 },
-  },
-
-  // Солодощі та десерти (розширено)
-  {
-    name: "Чизкейк",
-    per100: { kcal: 321, protein_g: 5.5, fat_g: 23, carbs_g: 26 },
-  },
-  {
-    name: "Тірамісу",
-    per100: { kcal: 240, protein_g: 5, fat_g: 14, carbs_g: 23 },
-  },
-  {
-    name: "Торт Наполеон",
-    per100: { kcal: 400, protein_g: 5, fat_g: 26, carbs_g: 38 },
-  },
-  {
-    name: "Торт Медовик",
-    per100: { kcal: 370, protein_g: 5, fat_g: 18, carbs_g: 48 },
-  },
-  {
-    name: "Торт Прага",
-    per100: { kcal: 387, protein_g: 4.5, fat_g: 21, carbs_g: 45 },
-  },
-  {
-    name: "Брауні",
-    per100: { kcal: 466, protein_g: 5, fat_g: 26, carbs_g: 55 },
-  },
-  {
-    name: "Молоко згущене",
-    per100: { kcal: 321, protein_g: 7.2, fat_g: 8.5, carbs_g: 56 },
-  },
-  {
-    name: "Халва",
-    per100: { kcal: 524, protein_g: 12, fat_g: 30, carbs_g: 54 },
-  },
-  {
-    name: "Курага в шоколаді",
-    per100: { kcal: 383, protein_g: 3.5, fat_g: 14, carbs_g: 64 },
-  },
-  {
-    name: "Карамель",
-    per100: { kcal: 385, protein_g: 0, fat_g: 0.1, carbs_g: 96 },
-  },
-  {
-    name: "Протеїновий батончик",
-    per100: { kcal: 380, protein_g: 30, fat_g: 14, carbs_g: 35 },
-  },
-  {
-    name: "Гранола батончик",
-    per100: { kcal: 400, protein_g: 6, fat_g: 14, carbs_g: 63 },
-  },
-
-  // Спортивне харчування (розширено)
-  {
-    name: "BCAA",
-    per100: { kcal: 340, protein_g: 85, fat_g: 0, carbs_g: 0 },
-  },
-  {
-    name: "Креатин",
-    per100: { kcal: 0, protein_g: 0, fat_g: 0, carbs_g: 0 },
-  },
-  {
-    name: "Ізотонік",
-    per100: { kcal: 24, protein_g: 0, fat_g: 0, carbs_g: 6 },
-  },
-
-  // Заморожені / готові снеки
-  {
-    name: "Млинці заморожені",
-    per100: { kcal: 207, protein_g: 5.5, fat_g: 7, carbs_g: 30 },
-  },
-  {
-    name: "Пельмені заморожені",
-    per100: { kcal: 248, protein_g: 13, fat_g: 10, carbs_g: 28 },
-  },
-  {
-    name: "Вареники заморожені",
-    per100: { kcal: 176, protein_g: 7, fat_g: 2, carbs_g: 33 },
-  },
-  {
-    name: "Котлета рибна",
-    per100: { kcal: 168, protein_g: 12, fat_g: 10, carbs_g: 8 },
-  },
-  {
-    name: "Мюслі батончик",
-    per100: { kcal: 438, protein_g: 7.5, fat_g: 18, carbs_g: 64 },
-  },
+  ...MEAT_AND_POULTRY,
+  ...FISH_AND_SEAFOOD,
+  ...EGGS,
+  ...DAIRY,
+  ...GRAINS,
+  ...BAKERY,
+  ...LEGUMES,
+  ...VEGETABLES,
+  ...FRUITS,
+  ...NUTS_AND_SEEDS,
+  ...OILS_AND_FATS,
+  ...SAUCES_AND_SPICES,
+  ...SWEETS,
+  ...BEVERAGES,
+  ...SPORTS_NUTRITION,
+  ...READY_MEALS,
+  ...UKRAINIAN_CUISINE,
+  ...SALADS,
+  ...FROZEN_SNACKS,
 ];
