+++ apps/web/src/modules/nutrition/lib/foodDb/seeds/bakery.ts
@@ -1,0 +1,80 @@
+import type { SeedFood } from "./types";
+
+export const BAKERY: SeedFood[] = [
+  {
+    name: "Хліб пшеничний",
+    per100: { kcal: 265, protein_g: 9, fat_g: 3.2, carbs_g: 49 },
+  },
+  {
+    name: "Хліб житній",
+    per100: { kcal: 259, protein_g: 9, fat_g: 3.3, carbs_g: 48 },
+  },
+  {
+    name: "Хліб цільнозерновий",
+    per100: { kcal: 247, protein_g: 9, fat_g: 3, carbs_g: 45 },
+  },
+  {
+    name: "Батон",
+    per100: { kcal: 262, protein_g: 8, fat_g: 3.2, carbs_g: 50 },
+  },
+  {
+    name: "Лаваш",
+    per100: { kcal: 277, protein_g: 9, fat_g: 1.2, carbs_g: 58 },
+  },
+  {
+    name: "Сухарі",
+    per100: { kcal: 347, protein_g: 11, fat_g: 4, carbs_g: 70 },
+  },
+  {
+    name: "Хлібці рисові",
+    per100: { kcal: 390, protein_g: 8, fat_g: 3, carbs_g: 84 },
+  },
+  {
+    name: "Багет",
+    per100: { kcal: 274, protein_g: 9, fat_g: 1.3, carbs_g: 56 },
+  },
+  {
+    name: "Чіабата",
+    per100: { kcal: 271, protein_g: 8.9, fat_g: 3.5, carbs_g: 51 },
+  },
+  {
+    name: "Тортилья пшенична",
+    per100: { kcal: 310, protein_g: 8, fat_g: 8, carbs_g: 51 },
+  },
+  {
+    name: "Хлібець гречаний",
+    per100: { kcal: 320, protein_g: 11, fat_g: 2.2, carbs_g: 63 },
+  },
+  {
+    name: "Булочка для бургера",
+    per100: { kcal: 279, protein_g: 9, fat_g: 5, carbs_g: 50 },
+  },
+  {
+    name: "Круасан",
+    per100: { kcal: 406, protein_g: 8.2, fat_g: 21, carbs_g: 46 },
+  },
+  {
+    name: "Булочка здобна",
+    per100: { kcal: 339, protein_g: 7.8, fat_g: 9.2, carbs_g: 56 },
+  },
+  {
+    name: "Пиріжок з капустою",
+    per100: { kcal: 223, protein_g: 5, fat_g: 6, carbs_g: 36 },
+  },
+  {
+    name: "Пиріжок з м'ясом",
+    per100: { kcal: 253, protein_g: 10, fat_g: 9, carbs_g: 32 },
+  },
+  {
+    name: "Маффін",
+    per100: { kcal: 377, protein_g: 5.5, fat_g: 15, carbs_g: 55 },
+  },
+  {
+    name: "Пончик",
+    per100: { kcal: 452, protein_g: 4.9, fat_g: 25, carbs_g: 51 },
+  },
+  {
+    name: "Еклер",
+    per100: { kcal: 262, protein_g: 6.3, fat_g: 16, carbs_g: 24 },
+  },
+];
