+++ apps/web/src/modules/nutrition/lib/foodDb/seeds/beverages.ts
@@ -1,0 +1,100 @@
+import type { SeedFood } from "./types";
+
+export const BEVERAGES: SeedFood[] = [
+  {
+    name: "Кава (без цукру)",
+    per100: { kcal: 1, protein_g: 0.1, fat_g: 0, carbs_g: 0 },
+  },
+  {
+    name: "Чай (без цукру)",
+    per100: { kcal: 1, protein_g: 0, fat_g: 0, carbs_g: 0 },
+  },
+  {
+    name: "Апельсиновий сік",
+    per100: { kcal: 45, protein_g: 0.7, fat_g: 0.2, carbs_g: 10 },
+  },
+  {
+    name: "Яблучний сік",
+    per100: { kcal: 46, protein_g: 0.1, fat_g: 0.1, carbs_g: 11 },
+  },
+  {
+    name: "Лате",
+    per100: { kcal: 56, protein_g: 3, fat_g: 3, carbs_g: 4.4 },
+  },
+  {
+    name: "Капучіно",
+    per100: { kcal: 37, protein_g: 2, fat_g: 2, carbs_g: 3 },
+  },
+  {
+    name: "Американо",
+    per100: { kcal: 2, protein_g: 0.1, fat_g: 0, carbs_g: 0 },
+  },
+  {
+    name: "Какао з молоком",
+    per100: { kcal: 76, protein_g: 3, fat_g: 3.5, carbs_g: 8 },
+  },
+  {
+    name: "Матча (лате)",
+    per100: { kcal: 55, protein_g: 2.5, fat_g: 2.5, carbs_g: 5.5 },
+  },
+  {
+    name: "Кола",
+    per100: { kcal: 42, protein_g: 0, fat_g: 0, carbs_g: 10.6 },
+  },
+  {
+    name: "Кола Zero",
+    per100: { kcal: 0.3, protein_g: 0, fat_g: 0, carbs_g: 0 },
+  },
+  {
+    name: "Спрайт",
+    per100: { kcal: 40, protein_g: 0, fat_g: 0, carbs_g: 10 },
+  },
+  {
+    name: "Фанта",
+    per100: { kcal: 48, protein_g: 0, fat_g: 0, carbs_g: 12 },
+  },
+  {
+    name: "Енергетик Red Bull",
+    per100: { kcal: 46, protein_g: 0.4, fat_g: 0, carbs_g: 11 },
+  },
+  {
+    name: "Квас",
+    per100: { kcal: 27, protein_g: 0.2, fat_g: 0, carbs_g: 5.2 },
+  },
+  {
+    name: "Узвар",
+    per100: { kcal: 48, protein_g: 0.3, fat_g: 0, carbs_g: 12 },
+  },
+  {
+    name: "Компот (без цукру)",
+    per100: { kcal: 24, protein_g: 0.1, fat_g: 0, carbs_g: 6 },
+  },
+  {
+    name: "Пиво світле 4.5%",
+    per100: { kcal: 43, protein_g: 0.5, fat_g: 0, carbs_g: 3.6 },
+  },
+  {
+    name: "Пиво темне",
+    per100: { kcal: 58, protein_g: 0.5, fat_g: 0, carbs_g: 4 },
+  },
+  {
+    name: "Вино сухе",
+    per100: { kcal: 82, protein_g: 0.1, fat_g: 0, carbs_g: 2.6 },
+  },
+  {
+    name: "Вино напівсолодке",
+    per100: { kcal: 98, protein_g: 0.2, fat_g: 0, carbs_g: 5.5 },
+  },
+  {
+    name: "Шампанське",
+    per100: { kcal: 88, protein_g: 0.2, fat_g: 0, carbs_g: 3 },
+  },
+  {
+    name: "Горілка 40%",
+    per100: { kcal: 231, protein_g: 0, fat_g: 0, carbs_g: 0.1 },
+  },
+  {
+    name: "Віскі 40%",
+    per100: { kcal: 250, protein_g: 0, fat_g: 0, carbs_g: 0 },
+  },
+];
