+++ apps/web/src/modules/nutrition/lib/foodDb/seeds/dairy.ts
@@ -1,0 +1,180 @@
+import type { SeedFood } from "./types";
+
+export const DAIRY: SeedFood[] = [
+  {
+    name: "Молоко 2.5%",
+    per100: { kcal: 52, protein_g: 3, fat_g: 2.5, carbs_g: 4.7 },
+  },
+  {
+    name: "Молоко 1%",
+    per100: { kcal: 42, protein_g: 3.4, fat_g: 1, carbs_g: 5 },
+  },
+  {
+    name: "Молоко 3.2%",
+    per100: { kcal: 60, protein_g: 3, fat_g: 3.2, carbs_g: 4.7 },
+  },
+  {
+    name: "Кефір 2.5%",
+    per100: { kcal: 53, protein_g: 3, fat_g: 2.5, carbs_g: 4 },
+  },
+  {
+    name: "Ряжанка",
+    per100: { kcal: 85, protein_g: 3, fat_g: 4.2, carbs_g: 9 },
+  },
+  {
+    name: "Йогурт натуральний 2%",
+    per100: { kcal: 73, protein_g: 4.3, fat_g: 2, carbs_g: 8 },
+  },
+  {
+    name: "Сир кисломолочний 5%",
+    per100: { kcal: 121, protein_g: 17, fat_g: 5, carbs_g: 3 },
+  },
+  {
+    name: "Сир кисломолочний 0%",
+    per100: { kcal: 71, protein_g: 18, fat_g: 0.6, carbs_g: 3 },
+  },
+  {
+    name: "Сир твердий",
+    per100: { kcal: 360, protein_g: 25, fat_g: 28, carbs_g: 1 },
+  },
+  {
+    name: "Сир моцарела",
+    per100: { kcal: 280, protein_g: 18, fat_g: 22, carbs_g: 2 },
+  },
+  {
+    name: "Сир фета",
+    per100: { kcal: 264, protein_g: 14, fat_g: 21, carbs_g: 4 },
+  },
+  {
+    name: "Сир пармезан",
+    per100: { kcal: 431, protein_g: 38, fat_g: 29, carbs_g: 4 },
+  },
+  {
+    name: "Сир рікотта",
+    per100: { kcal: 174, protein_g: 11, fat_g: 13, carbs_g: 3 },
+  },
+  {
+    name: "Сметана 20%",
+    per100: { kcal: 204, protein_g: 2.8, fat_g: 20, carbs_g: 3.2 },
+  },
+  {
+    name: "Сметана 10%",
+    per100: { kcal: 115, protein_g: 3, fat_g: 10, carbs_g: 4 },
+  },
+  {
+    name: "Вершки 20%",
+    per100: { kcal: 207, protein_g: 2.5, fat_g: 20, carbs_g: 3.7 },
+  },
+  {
+    name: "Вершки 10%",
+    per100: { kcal: 119, protein_g: 2.7, fat_g: 10, carbs_g: 4.5 },
+  },
+  {
+    name: "Вершкове масло",
+    per100: { kcal: 717, protein_g: 0.9, fat_g: 81, carbs_g: 0.1 },
+  },
+  {
+    name: "Морозиво вершкове",
+    per100: { kcal: 207, protein_g: 3.5, fat_g: 11, carbs_g: 24 },
+  },
+  {
+    name: "Молоко безлактозне",
+    per100: { kcal: 46, protein_g: 3.3, fat_g: 1.5, carbs_g: 4.7 },
+  },
+  {
+    name: "Молоко знежирене",
+    per100: { kcal: 35, protein_g: 3.4, fat_g: 0.1, carbs_g: 5 },
+  },
+  {
+    name: "Молоко кокосове",
+    per100: { kcal: 230, protein_g: 2.3, fat_g: 24, carbs_g: 5.5 },
+  },
+  {
+    name: "Молоко мигдалеве",
+    per100: { kcal: 17, protein_g: 0.6, fat_g: 1.1, carbs_g: 1.5 },
+  },
+  {
+    name: "Молоко вівсяне",
+    per100: { kcal: 45, protein_g: 0.8, fat_g: 1.5, carbs_g: 6.7 },
+  },
+  {
+    name: "Молоко соєве",
+    per100: { kcal: 45, protein_g: 3.3, fat_g: 1.8, carbs_g: 3.3 },
+  },
+  {
+    name: "Йогурт грецький 2%",
+    per100: { kcal: 73, protein_g: 10, fat_g: 2, carbs_g: 3.6 },
+  },
+  {
+    name: "Йогурт грецький 0%",
+    per100: { kcal: 59, protein_g: 10, fat_g: 0.4, carbs_g: 3.6 },
+  },
+  {
+    name: "Йогурт питний",
+    per100: { kcal: 72, protein_g: 2.8, fat_g: 1.5, carbs_g: 12 },
+  },
+  {
+    name: "Сир кисломолочний 9%",
+    per100: { kcal: 159, protein_g: 17, fat_g: 9, carbs_g: 2 },
+  },
+  {
+    name: "Сир кисломолочний 18%",
+    per100: { kcal: 236, protein_g: 14, fat_g: 18, carbs_g: 2.8 },
+  },
+  {
+    name: "Сирок глазурований",
+    per100: { kcal: 407, protein_g: 8.5, fat_g: 27, carbs_g: 32 },
+  },
+  {
+    name: "Сир чеддер",
+    per100: { kcal: 403, protein_g: 25, fat_g: 33, carbs_g: 1.3 },
+  },
+  {
+    name: "Сир гауда",
+    per100: { kcal: 356, protein_g: 25, fat_g: 27, carbs_g: 2.2 },
+  },
+  {
+    name: "Сир камамбер",
+    per100: { kcal: 299, protein_g: 20, fat_g: 24, carbs_g: 0.5 },
+  },
+  {
+    name: "Сир брі",
+    per100: { kcal: 334, protein_g: 21, fat_g: 28, carbs_g: 0.5 },
+  },
+  {
+    name: "Сир дор блю",
+    per100: { kcal: 353, protein_g: 21, fat_g: 29, carbs_g: 2.3 },
+  },
+  {
+    name: "Сир бринза",
+    per100: { kcal: 262, protein_g: 18, fat_g: 20, carbs_g: 2 },
+  },
+  {
+    name: "Сир сулугуні",
+    per100: { kcal: 290, protein_g: 20, fat_g: 22, carbs_g: 0.4 },
+  },
+  {
+    name: "Сир вершковий",
+    per100: { kcal: 342, protein_g: 6, fat_g: 34, carbs_g: 4 },
+  },
+  {
+    name: "Маскарпоне",
+    per100: { kcal: 430, protein_g: 5, fat_g: 45, carbs_g: 4.5 },
+  },
+  {
+    name: "Сметана 15%",
+    per100: { kcal: 158, protein_g: 2.8, fat_g: 15, carbs_g: 3.6 },
+  },
+  {
+    name: "Сметана 30%",
+    per100: { kcal: 293, protein_g: 2.4, fat_g: 30, carbs_g: 3.1 },
+  },
+  {
+    name: "Ацидофілін",
+    per100: { kcal: 57, protein_g: 2.8, fat_g: 3.2, carbs_g: 3.8 },
+  },
+  {
+    name: "Сироватка",
+    per100: { kcal: 20, protein_g: 0.8, fat_g: 0.2, carbs_g: 3.5 },
+  },
+];
