+++ apps/web/src/modules/nutrition/lib/foodDb/seeds/eggs.ts
@@ -1,0 +1,32 @@
+import type { SeedFood } from "./types";
+
+export const EGGS: SeedFood[] = [
+  {
+    name: "Яйце куряче",
+    per100: { kcal: 143, protein_g: 13, fat_g: 10, carbs_g: 1.1 },
+  },
+  {
+    name: "Яйце варене",
+    per100: { kcal: 155, protein_g: 13, fat_g: 11, carbs_g: 1.1 },
+  },
+  {
+    name: "Яєчня",
+    per100: { kcal: 196, protein_g: 14, fat_g: 15, carbs_g: 1 },
+  },
+  {
+    name: "Омлет",
+    per100: { kcal: 154, protein_g: 10, fat_g: 12, carbs_g: 3 },
+  },
+  {
+    name: "Яйце перепелине",
+    per100: { kcal: 158, protein_g: 13, fat_g: 11, carbs_g: 0.6 },
+  },
+  {
+    name: "Яєчний білок",
+    per100: { kcal: 52, protein_g: 11, fat_g: 0.2, carbs_g: 0.7 },
+  },
+  {
+    name: "Яєчний жовток",
+    per100: { kcal: 322, protein_g: 16, fat_g: 27, carbs_g: 3.6 },
+  },
+];
