+++ apps/web/src/modules/nutrition/lib/foodDb/seeds/fishAndSeafood.ts
@@ -1,0 +1,98 @@
+import type { SeedFood } from "./types";
+
+export const FISH_AND_SEAFOOD: SeedFood[] = [
+  {
+    name: "Лосось",
+    per100: { kcal: 208, protein_g: 20, fat_g: 13, carbs_g: 0 },
+  },
+  {
+    name: "Тунець (у власному соку)",
+    per100: { kcal: 116, protein_g: 26, fat_g: 1, carbs_g: 0 },
+  },
+  {
+    name: "Оселедець",
+    per100: { kcal: 218, protein_g: 18, fat_g: 16, carbs_g: 0 },
+  },
+  {
+    name: "Скумбрія",
+    per100: { kcal: 205, protein_g: 19, fat_g: 14, carbs_g: 0 },
+  },
+  {
+    name: "Тріска",
+    per100: { kcal: 82, protein_g: 18, fat_g: 0.7, carbs_g: 0 },
+  },
+  {
+    name: "Форель",
+    per100: { kcal: 148, protein_g: 21, fat_g: 7, carbs_g: 0 },
+  },
+  {
+    name: "Горбуша",
+    per100: { kcal: 142, protein_g: 21, fat_g: 6, carbs_g: 0 },
+  },
+  { name: "Минтай", per100: { kcal: 82, protein_g: 18, fat_g: 1, carbs_g: 0 } },
+  { name: "Короп", per100: { kcal: 127, protein_g: 18, fat_g: 6, carbs_g: 0 } },
+  {
+    name: "Тілапія",
+    per100: { kcal: 128, protein_g: 26, fat_g: 3, carbs_g: 0 },
+  },
+  {
+    name: "Кальмар",
+    per100: { kcal: 92, protein_g: 18, fat_g: 1.4, carbs_g: 1.9 },
+  },
+  {
+    name: "Креветки",
+    per100: { kcal: 99, protein_g: 19, fat_g: 1.8, carbs_g: 0.9 },
+  },
+  {
+    name: "Мідії",
+    per100: { kcal: 86, protein_g: 12, fat_g: 2.2, carbs_g: 3.7 },
+  },
+  {
+    name: "Ікра червона",
+    per100: { kcal: 263, protein_g: 32, fat_g: 15, carbs_g: 0 },
+  },
+  {
+    name: "Ікра чорна",
+    per100: { kcal: 264, protein_g: 25, fat_g: 18, carbs_g: 0 },
+  },
+  {
+    name: "Крабові палички",
+    per100: { kcal: 88, protein_g: 7, fat_g: 1, carbs_g: 14 },
+  },
+  {
+    name: "Оселедець солений",
+    per100: { kcal: 217, protein_g: 19, fat_g: 15, carbs_g: 0 },
+  },
+  {
+    name: "Шпроти в олії",
+    per100: { kcal: 363, protein_g: 17, fat_g: 32, carbs_g: 0.4 },
+  },
+  {
+    name: "Сардина в олії",
+    per100: { kcal: 208, protein_g: 25, fat_g: 11, carbs_g: 0 },
+  },
+  {
+    name: "Судак",
+    per100: { kcal: 84, protein_g: 19, fat_g: 1.1, carbs_g: 0 },
+  },
+  {
+    name: "Сом",
+    per100: { kcal: 115, protein_g: 16, fat_g: 5.3, carbs_g: 0 },
+  },
+  {
+    name: "Хек",
+    per100: { kcal: 86, protein_g: 16, fat_g: 2.2, carbs_g: 0 },
+  },
+  {
+    name: "Оселедець маринований",
+    per100: { kcal: 262, protein_g: 17, fat_g: 21, carbs_g: 1 },
+  },
+  {
+    name: "Устриці",
+    per100: { kcal: 68, protein_g: 9, fat_g: 2.3, carbs_g: 3.9 },
+  },
+  {
+    name: "Восьминіг",
+    per100: { kcal: 82, protein_g: 15, fat_g: 1, carbs_g: 2.2 },
+  },
+];
