+++ apps/web/src/modules/nutrition/lib/foodDb/seeds/frozenSnacks.ts
@@ -1,0 +1,24 @@
+import type { SeedFood } from "./types";
+
+export const FROZEN_SNACKS: SeedFood[] = [
+  {
+    name: "Млинці заморожені",
+    per100: { kcal: 207, protein_g: 5.5, fat_g: 7, carbs_g: 30 },
+  },
+  {
+    name: "Пельмені заморожені",
+    per100: { kcal: 248, protein_g: 13, fat_g: 10, carbs_g: 28 },
+  },
+  {
+    name: "Вареники заморожені",
+    per100: { kcal: 176, protein_g: 7, fat_g: 2, carbs_g: 33 },
+  },
+  {
+    name: "Котлета рибна",
+    per100: { kcal: 168, protein_g: 12, fat_g: 10, carbs_g: 8 },
+  },
+  {
+    name: "Мюслі батончик",
+    per100: { kcal: 438, protein_g: 7.5, fat_g: 18, carbs_g: 64 },
+  },
+];
