+++ apps/web/src/modules/nutrition/lib/foodDb/seeds/fruits.ts
@@ -1,0 +1,160 @@
+import type { SeedFood } from "./types";
+
+export const FRUITS: SeedFood[] = [
+  {
+    name: "Яблуко",
+    per100: { kcal: 52, protein_g: 0.3, fat_g: 0.2, carbs_g: 14 },
+  },
+  {
+    name: "Груша",
+    per100: { kcal: 57, protein_g: 0.4, fat_g: 0.1, carbs_g: 15 },
+  },
+  {
+    name: "Банан",
+    per100: { kcal: 89, protein_g: 1.1, fat_g: 0.3, carbs_g: 23 },
+  },
+  {
+    name: "Апельсин",
+    per100: { kcal: 47, protein_g: 0.9, fat_g: 0.1, carbs_g: 12 },
+  },
+  {
+    name: "Мандарин",
+    per100: { kcal: 53, protein_g: 0.8, fat_g: 0.3, carbs_g: 13 },
+  },
+  {
+    name: "Лимон",
+    per100: { kcal: 29, protein_g: 1.1, fat_g: 0.3, carbs_g: 9 },
+  },
+  {
+    name: "Грейпфрут",
+    per100: { kcal: 42, protein_g: 0.8, fat_g: 0.1, carbs_g: 11 },
+  },
+  {
+    name: "Ківі",
+    per100: { kcal: 61, protein_g: 1.1, fat_g: 0.5, carbs_g: 14 },
+  },
+  {
+    name: "Манго",
+    per100: { kcal: 60, protein_g: 0.8, fat_g: 0.4, carbs_g: 15 },
+  },
+  {
+    name: "Ананас",
+    per100: { kcal: 50, protein_g: 0.5, fat_g: 0.1, carbs_g: 13 },
+  },
+  {
+    name: "Персик",
+    per100: { kcal: 39, protein_g: 0.9, fat_g: 0.3, carbs_g: 9.5 },
+  },
+  {
+    name: "Абрикос",
+    per100: { kcal: 48, protein_g: 1.4, fat_g: 0.4, carbs_g: 11 },
+  },
+  {
+    name: "Слива",
+    per100: { kcal: 46, protein_g: 0.7, fat_g: 0.3, carbs_g: 11 },
+  },
+  {
+    name: "Вишня",
+    per100: { kcal: 63, protein_g: 1, fat_g: 0.3, carbs_g: 16 },
+  },
+  {
+    name: "Черешня",
+    per100: { kcal: 63, protein_g: 1.1, fat_g: 0.4, carbs_g: 16 },
+  },
+  {
+    name: "Виноград",
+    per100: { kcal: 69, protein_g: 0.7, fat_g: 0.2, carbs_g: 18 },
+  },
+  {
+    name: "Полуниця",
+    per100: { kcal: 32, protein_g: 0.7, fat_g: 0.3, carbs_g: 7.7 },
+  },
+  {
+    name: "Малина",
+    per100: { kcal: 52, protein_g: 1.2, fat_g: 0.7, carbs_g: 12 },
+  },
+  {
+    name: "Чорниця",
+    per100: { kcal: 57, protein_g: 0.7, fat_g: 0.3, carbs_g: 14 },
+  },
+  {
+    name: "Смородина чорна",
+    per100: { kcal: 63, protein_g: 1.4, fat_g: 0.4, carbs_g: 15 },
+  },
+  {
+    name: "Смородина червона",
+    per100: { kcal: 56, protein_g: 1.4, fat_g: 0.2, carbs_g: 13 },
+  },
+  {
+    name: "Журавлина",
+    per100: { kcal: 46, protein_g: 0.4, fat_g: 0.1, carbs_g: 12 },
+  },
+  {
+    name: "Кавун",
+    per100: { kcal: 30, protein_g: 0.6, fat_g: 0.2, carbs_g: 7.5 },
+  },
+  {
+    name: "Диня",
+    per100: { kcal: 34, protein_g: 0.8, fat_g: 0.2, carbs_g: 8 },
+  },
+  {
+    name: "Хурма",
+    per100: { kcal: 70, protein_g: 0.6, fat_g: 0.4, carbs_g: 19 },
+  },
+  {
+    name: "Гранат",
+    per100: { kcal: 83, protein_g: 1.7, fat_g: 1.2, carbs_g: 19 },
+  },
+  {
+    name: "Інжир",
+    per100: { kcal: 74, protein_g: 0.8, fat_g: 0.3, carbs_g: 19 },
+  },
+  {
+    name: "Чорнослив",
+    per100: { kcal: 240, protein_g: 2.2, fat_g: 0.4, carbs_g: 64 },
+  },
+  {
+    name: "Родзинки",
+    per100: { kcal: 299, protein_g: 3.1, fat_g: 0.5, carbs_g: 79 },
+  },
+  {
+    name: "Кокос",
+    per100: { kcal: 354, protein_g: 3.3, fat_g: 33, carbs_g: 15 },
+  },
+  {
+    name: "Фініки",
+    per100: { kcal: 282, protein_g: 2.5, fat_g: 0.4, carbs_g: 75 },
+  },
+  {
+    name: "Курага",
+    per100: { kcal: 241, protein_g: 3.4, fat_g: 0.5, carbs_g: 63 },
+  },
+  {
+    name: "Фейхоа",
+    per100: { kcal: 49, protein_g: 0.7, fat_g: 0.6, carbs_g: 11 },
+  },
+  {
+    name: "Нектарин",
+    per100: { kcal: 44, protein_g: 1.1, fat_g: 0.3, carbs_g: 11 },
+  },
+  {
+    name: "Папайя",
+    per100: { kcal: 43, protein_g: 0.5, fat_g: 0.3, carbs_g: 11 },
+  },
+  {
+    name: "Агрус",
+    per100: { kcal: 44, protein_g: 0.9, fat_g: 0.6, carbs_g: 9.8 },
+  },
+  {
+    name: "Обліпиха",
+    per100: { kcal: 82, protein_g: 1.2, fat_g: 5.4, carbs_g: 5.7 },
+  },
+  {
+    name: "Лохина",
+    per100: { kcal: 57, protein_g: 0.7, fat_g: 0.3, carbs_g: 14 },
+  },
+  {
+    name: "Калина",
+    per100: { kcal: 26, protein_g: 0.4, fat_g: 0, carbs_g: 7 },
+  },
+];
