+++ apps/web/src/modules/nutrition/lib/foodDb/seeds/grains.ts
@@ -1,0 +1,109 @@
+import type { SeedFood } from "./types";
+
+export const GRAINS: SeedFood[] = [
+  {
+    name: "Вівсянка (суха)",
+    per100: { kcal: 389, protein_g: 17, fat_g: 7, carbs_g: 66 },
+  },
+  {
+    name: "Вівсянка варена",
+    per100: { kcal: 71, protein_g: 2.5, fat_g: 1.4, carbs_g: 12 },
+  },
+  {
+    name: "Рис (сухий)",
+    per100: { kcal: 365, protein_g: 7, fat_g: 0.7, carbs_g: 80 },
+  },
+  {
+    name: "Рис варений",
+    per100: { kcal: 130, protein_g: 2.4, fat_g: 0.3, carbs_g: 28 },
+  },
+  {
+    name: "Гречка (суха)",
+    per100: { kcal: 343, protein_g: 13, fat_g: 3.4, carbs_g: 71 },
+  },
+  {
+    name: "Гречка варена",
+    per100: { kcal: 110, protein_g: 3.6, fat_g: 1.2, carbs_g: 20 },
+  },
+  {
+    name: "Макарони (сухі)",
+    per100: { kcal: 371, protein_g: 13, fat_g: 1.5, carbs_g: 75 },
+  },
+  {
+    name: "Паста варена",
+    per100: { kcal: 158, protein_g: 5.8, fat_g: 0.9, carbs_g: 31 },
+  },
+  {
+    name: "Булгур (сухий)",
+    per100: { kcal: 342, protein_g: 12, fat_g: 1.3, carbs_g: 76 },
+  },
+  {
+    name: "Пшоно (сухе)",
+    per100: { kcal: 378, protein_g: 11, fat_g: 3.9, carbs_g: 73 },
+  },
+  {
+    name: "Перлова крупа (суха)",
+    per100: { kcal: 352, protein_g: 9.9, fat_g: 1.2, carbs_g: 73 },
+  },
+  {
+    name: "Кіноа (суха)",
+    per100: { kcal: 368, protein_g: 14, fat_g: 6.1, carbs_g: 64 },
+  },
+  {
+    name: "Кус-кус (сухий)",
+    per100: { kcal: 376, protein_g: 13, fat_g: 0.6, carbs_g: 77 },
+  },
+  {
+    name: "Манна крупа",
+    per100: { kcal: 360, protein_g: 10, fat_g: 1, carbs_g: 73 },
+  },
+  {
+    name: "Кукурудзяна крупа",
+    per100: { kcal: 337, protein_g: 8, fat_g: 1.2, carbs_g: 75 },
+  },
+  {
+    name: "Вівсяні пластівці швидкі",
+    per100: { kcal: 352, protein_g: 11, fat_g: 6, carbs_g: 66 },
+  },
+  { name: "Мюслі", per100: { kcal: 367, protein_g: 9, fat_g: 6, carbs_g: 70 } },
+  {
+    name: "Гранола",
+    per100: { kcal: 471, protein_g: 9, fat_g: 16, carbs_g: 75 },
+  },
+  {
+    name: "Кукурудзяні пластівці",
+    per100: { kcal: 370, protein_g: 7, fat_g: 1, carbs_g: 84 },
+  },
+  {
+    name: "Пшоно варене",
+    per100: { kcal: 90, protein_g: 3, fat_g: 0.7, carbs_g: 17 },
+  },
+  {
+    name: "Булгур варений",
+    per100: { kcal: 83, protein_g: 3.1, fat_g: 0.2, carbs_g: 19 },
+  },
+  {
+    name: "Кіноа варена",
+    per100: { kcal: 120, protein_g: 4.4, fat_g: 1.9, carbs_g: 21 },
+  },
+  {
+    name: "Кус-кус варений",
+    per100: { kcal: 112, protein_g: 3.8, fat_g: 0.2, carbs_g: 23 },
+  },
+  {
+    name: "Перлова варена",
+    per100: { kcal: 123, protein_g: 2.3, fat_g: 0.4, carbs_g: 28 },
+  },
+  {
+    name: "Рис бурий варений",
+    per100: { kcal: 112, protein_g: 2.6, fat_g: 0.9, carbs_g: 23 },
+  },
+  {
+    name: "Амарант",
+    per100: { kcal: 371, protein_g: 14, fat_g: 7, carbs_g: 65 },
+  },
+  {
+    name: "Полента",
+    per100: { kcal: 70, protein_g: 1.5, fat_g: 0.3, carbs_g: 15 },
+  },
+];
