+++ apps/web/src/modules/nutrition/lib/foodDb/seeds/legumes.ts
@@ -1,0 +1,28 @@
+import type { SeedFood } from "./types";
+
+export const LEGUMES: SeedFood[] = [
+  {
+    name: "Квасоля (варена)",
+    per100: { kcal: 127, protein_g: 9, fat_g: 0.5, carbs_g: 23 },
+  },
+  {
+    name: "Сочевиця (варена)",
+    per100: { kcal: 116, protein_g: 9, fat_g: 0.4, carbs_g: 20 },
+  },
+  {
+    name: "Нут (варений)",
+    per100: { kcal: 164, protein_g: 9, fat_g: 2.6, carbs_g: 27 },
+  },
+  {
+    name: "Тофу",
+    per100: { kcal: 76, protein_g: 8, fat_g: 4.8, carbs_g: 1.9 },
+  },
+  {
+    name: "Едамаме",
+    per100: { kcal: 121, protein_g: 11, fat_g: 5.2, carbs_g: 8.9 },
+  },
+  {
+    name: "Горошок зелений",
+    per100: { kcal: 81, protein_g: 5, fat_g: 0.4, carbs_g: 14 },
+  },
+];
