+++ apps/web/src/modules/nutrition/lib/foodDb/seeds/meatAndPoultry.ts
@@ -1,0 +1,156 @@
+import type { SeedFood } from "./types";
+
+export const MEAT_AND_POULTRY: SeedFood[] = [
+  {
+    name: "Куряча грудка",
+    per100: { kcal: 165, protein_g: 31, fat_g: 3.6, carbs_g: 0 },
+  },
+  {
+    name: "Куряче стегно",
+    per100: { kcal: 215, protein_g: 18, fat_g: 15, carbs_g: 0 },
+  },
+  {
+    name: "Куряче крило",
+    per100: { kcal: 222, protein_g: 18, fat_g: 15, carbs_g: 0 },
+  },
+  {
+    name: "Індичка (філе)",
+    per100: { kcal: 135, protein_g: 29, fat_g: 1.6, carbs_g: 0 },
+  },
+  {
+    name: "Яловичина",
+    per100: { kcal: 187, protein_g: 20, fat_g: 12, carbs_g: 0 },
+  },
+  {
+    name: "Яловичий фарш",
+    per100: { kcal: 254, protein_g: 17, fat_g: 20, carbs_g: 0 },
+  },
+  {
+    name: "Яловичий стейк",
+    per100: { kcal: 271, protein_g: 26, fat_g: 18, carbs_g: 0 },
+  },
+  {
+    name: "Свинина",
+    per100: { kcal: 242, protein_g: 17, fat_g: 19, carbs_g: 0 },
+  },
+  {
+    name: "Свинячий фарш",
+    per100: { kcal: 263, protein_g: 16, fat_g: 22, carbs_g: 0 },
+  },
+  {
+    name: "Курячий фарш",
+    per100: { kcal: 143, protein_g: 17, fat_g: 8, carbs_g: 0 },
+  },
+  {
+    name: "Шинка",
+    per100: { kcal: 163, protein_g: 17, fat_g: 10, carbs_g: 1.5 },
+  },
+  {
+    name: "Бекон",
+    per100: { kcal: 541, protein_g: 37, fat_g: 42, carbs_g: 1.5 },
+  },
+  {
+    name: "Ковбаса варена",
+    per100: { kcal: 260, protein_g: 12, fat_g: 23, carbs_g: 1 },
+  },
+  {
+    name: "Ковбаса копчена",
+    per100: { kcal: 420, protein_g: 17, fat_g: 38, carbs_g: 2 },
+  },
+  {
+    name: "Сосиски",
+    per100: { kcal: 256, protein_g: 10, fat_g: 23, carbs_g: 2 },
+  },
+  {
+    name: "Котлета м'ясна",
+    per100: { kcal: 230, protein_g: 15, fat_g: 18, carbs_g: 5 },
+  },
+  {
+    name: "Котлета куряча",
+    per100: { kcal: 200, protein_g: 17, fat_g: 14, carbs_g: 5 },
+  },
+  {
+    name: "Кролик",
+    per100: { kcal: 183, protein_g: 21, fat_g: 11, carbs_g: 0 },
+  },
+  {
+    name: "Качка",
+    per100: { kcal: 337, protein_g: 19, fat_g: 28, carbs_g: 0 },
+  },
+  {
+    name: "Шашлик свинячий",
+    per100: { kcal: 280, protein_g: 18, fat_g: 22, carbs_g: 2 },
+  },
+  {
+    name: "Сало",
+    per100: { kcal: 797, protein_g: 2.4, fat_g: 89, carbs_g: 0 },
+  },
+  {
+    name: "Хамон",
+    per100: { kcal: 241, protein_g: 31, fat_g: 13, carbs_g: 0 },
+  },
+  {
+    name: "Пармська шинка",
+    per100: { kcal: 250, protein_g: 26, fat_g: 16, carbs_g: 0.3 },
+  },
+  {
+    name: "Саламі",
+    per100: { kcal: 407, protein_g: 22, fat_g: 34, carbs_g: 1.6 },
+  },
+  {
+    name: "Пепероні",
+    per100: { kcal: 494, protein_g: 23, fat_g: 44, carbs_g: 1.2 },
+  },
+  {
+    name: "Бастурма",
+    per100: { kcal: 240, protein_g: 50, fat_g: 4, carbs_g: 0 },
+  },
+  {
+    name: "Холодець",
+    per100: { kcal: 120, protein_g: 12, fat_g: 8, carbs_g: 0.5 },
+  },
+  {
+    name: "Печінка куряча",
+    per100: { kcal: 137, protein_g: 17, fat_g: 6.5, carbs_g: 1.4 },
+  },
+  {
+    name: "Печінка яловича",
+    per100: { kcal: 135, protein_g: 20, fat_g: 3.6, carbs_g: 5.3 },
+  },
+  {
+    name: "Серце куряче",
+    per100: { kcal: 153, protein_g: 16, fat_g: 9.3, carbs_g: 0.7 },
+  },
+  {
+    name: "Шлунки курячі",
+    per100: { kcal: 94, protein_g: 18, fat_g: 2.1, carbs_g: 0 },
+  },
+  {
+    name: "Баранина",
+    per100: { kcal: 294, protein_g: 25, fat_g: 21, carbs_g: 0 },
+  },
+  {
+    name: "Гусятина",
+    per100: { kcal: 305, protein_g: 16, fat_g: 27, carbs_g: 0 },
+  },
+  {
+    name: "Телятина",
+    per100: { kcal: 172, protein_g: 20, fat_g: 10, carbs_g: 0 },
+  },
+  {
+    name: "Філе індички",
+    per100: { kcal: 112, protein_g: 24, fat_g: 1.7, carbs_g: 0 },
+  },
+  {
+    name: "Куряча ніжка",
+    per100: { kcal: 184, protein_g: 20, fat_g: 11, carbs_g: 0 },
+  },
+  {
+    name: "Котлета по-київськи",
+    per100: { kcal: 290, protein_g: 15, fat_g: 22, carbs_g: 9 },
+  },
+  {
+    name: "Нагетси курячі",
+    per100: { kcal: 297, protein_g: 15, fat_g: 19, carbs_g: 16 },
+  },
+];
