+++ apps/web/src/modules/nutrition/lib/foodDb/seeds/nutsAndSeeds.ts
@@ -1,0 +1,48 @@
+import type { SeedFood } from "./types";
+
+export const NUTS_AND_SEEDS: SeedFood[] = [
+  {
+    name: "Мигдаль",
+    per100: { kcal: 579, protein_g: 21, fat_g: 50, carbs_g: 22 },
+  },
+  {
+    name: "Горіхи волоські",
+    per100: { kcal: 654, protein_g: 15, fat_g: 65, carbs_g: 14 },
+  },
+  {
+    name: "Арахіс",
+    per100: { kcal: 567, protein_g: 26, fat_g: 49, carbs_g: 16 },
+  },
+  {
+    name: "Кешью",
+    per100: { kcal: 553, protein_g: 18, fat_g: 44, carbs_g: 30 },
+  },
+  {
+    name: "Фісташки",
+    per100: { kcal: 562, protein_g: 20, fat_g: 45, carbs_g: 28 },
+  },
+  {
+    name: "Фундук",
+    per100: { kcal: 628, protein_g: 15, fat_g: 61, carbs_g: 17 },
+  },
+  {
+    name: "Насіння соняшнику",
+    per100: { kcal: 584, protein_g: 21, fat_g: 51, carbs_g: 20 },
+  },
+  {
+    name: "Насіння гарбуза",
+    per100: { kcal: 559, protein_g: 30, fat_g: 49, carbs_g: 11 },
+  },
+  {
+    name: "Насіння льону",
+    per100: { kcal: 534, protein_g: 18, fat_g: 42, carbs_g: 29 },
+  },
+  {
+    name: "Кунжут",
+    per100: { kcal: 573, protein_g: 17, fat_g: 50, carbs_g: 23 },
+  },
+  {
+    name: "Паста арахісова",
+    per100: { kcal: 588, protein_g: 25, fat_g: 50, carbs_g: 20 },
+  },
+];
