+++ apps/web/src/modules/nutrition/lib/foodDb/seeds/oilsAndFats.ts
@@ -1,0 +1,16 @@
+import type { SeedFood } from "./types";
+
+export const OILS_AND_FATS: SeedFood[] = [
+  {
+    name: "Оливкова олія",
+    per100: { kcal: 884, protein_g: 0, fat_g: 100, carbs_g: 0 },
+  },
+  {
+    name: "Олія соняшникова",
+    per100: { kcal: 884, protein_g: 0, fat_g: 100, carbs_g: 0 },
+  },
+  {
+    name: "Олія кокосова",
+    per100: { kcal: 862, protein_g: 0, fat_g: 100, carbs_g: 0 },
+  },
+];
