+++ apps/web/src/modules/nutrition/lib/foodDb/seeds/readyMeals.ts
@@ -1,0 +1,98 @@
+import type { SeedFood } from "./types";
+
+export const READY_MEALS: SeedFood[] = [
+  {
+    name: "Вареники з картоплею",
+    per100: { kcal: 197, protein_g: 6, fat_g: 4, carbs_g: 34 },
+  },
+  {
+    name: "Вареники з сиром",
+    per100: { kcal: 222, protein_g: 9, fat_g: 5, carbs_g: 36 },
+  },
+  {
+    name: "Пельмені",
+    per100: { kcal: 235, protein_g: 12, fat_g: 10, carbs_g: 26 },
+  },
+  { name: "Борщ", per100: { kcal: 58, protein_g: 3, fat_g: 2.5, carbs_g: 7 } },
+  {
+    name: "Пюре картопляне",
+    per100: { kcal: 87, protein_g: 2, fat_g: 3, carbs_g: 14 },
+  },
+  { name: "Плов", per100: { kcal: 190, protein_g: 8, fat_g: 8, carbs_g: 24 } },
+  {
+    name: "Тушкована капуста",
+    per100: { kcal: 75, protein_g: 2, fat_g: 4, carbs_g: 9 },
+  },
+  {
+    name: "Піца",
+    per100: { kcal: 266, protein_g: 11, fat_g: 10, carbs_g: 33 },
+  },
+  {
+    name: "Суші/ролл",
+    per100: { kcal: 170, protein_g: 5, fat_g: 2.5, carbs_g: 32 },
+  },
+  {
+    name: "Бургер",
+    per100: { kcal: 295, protein_g: 17, fat_g: 14, carbs_g: 27 },
+  },
+  {
+    name: "Чізбургер",
+    per100: { kcal: 303, protein_g: 16, fat_g: 15, carbs_g: 26 },
+  },
+  {
+    name: "Хот-дог",
+    per100: { kcal: 247, protein_g: 10, fat_g: 15, carbs_g: 18 },
+  },
+  {
+    name: "Шаурма",
+    per100: { kcal: 215, protein_g: 12, fat_g: 11, carbs_g: 18 },
+  },
+  {
+    name: "Картопля фрі",
+    per100: { kcal: 312, protein_g: 3.4, fat_g: 15, carbs_g: 41 },
+  },
+  {
+    name: "Картопля по-селянськи",
+    per100: { kcal: 147, protein_g: 2.5, fat_g: 7, carbs_g: 19 },
+  },
+  {
+    name: "Сендвіч з куркою",
+    per100: { kcal: 221, protein_g: 13, fat_g: 9, carbs_g: 22 },
+  },
+  {
+    name: "Ролл з лососем",
+    per100: { kcal: 184, protein_g: 7, fat_g: 4, carbs_g: 30 },
+  },
+  {
+    name: "Піца Маргарита",
+    per100: { kcal: 251, protein_g: 11, fat_g: 8, carbs_g: 33 },
+  },
+  {
+    name: "Піца Пепероні",
+    per100: { kcal: 296, protein_g: 12, fat_g: 13, carbs_g: 32 },
+  },
+  {
+    name: "Піца 4 сири",
+    per100: { kcal: 313, protein_g: 14, fat_g: 15, carbs_g: 30 },
+  },
+  {
+    name: "Паста карбонара",
+    per100: { kcal: 200, protein_g: 9, fat_g: 10, carbs_g: 20 },
+  },
+  {
+    name: "Паста болоньєзе",
+    per100: { kcal: 160, protein_g: 8, fat_g: 5, carbs_g: 21 },
+  },
+  {
+    name: "Лазанья",
+    per100: { kcal: 135, protein_g: 8, fat_g: 5, carbs_g: 14 },
+  },
+  {
+    name: "Рагу овочеве",
+    per100: { kcal: 56, protein_g: 1.8, fat_g: 2, carbs_g: 8 },
+  },
+  {
+    name: "Тушкована яловичина",
+    per100: { kcal: 232, protein_g: 16, fat_g: 17, carbs_g: 3 },
+  },
+];
