+++ apps/web/src/modules/nutrition/lib/foodDb/seeds/salads.ts
@@ -1,0 +1,36 @@
+import type { SeedFood } from "./types";
+
+export const SALADS: SeedFood[] = [
+  {
+    name: "Олів'є",
+    per100: { kcal: 198, protein_g: 5, fat_g: 14, carbs_g: 12 },
+  },
+  {
+    name: "Вінегрет",
+    per100: { kcal: 76, protein_g: 1.5, fat_g: 4, carbs_g: 8 },
+  },
+  {
+    name: "Оселедець під шубою",
+    per100: { kcal: 193, protein_g: 5, fat_g: 15, carbs_g: 9 },
+  },
+  {
+    name: "Цезар з куркою",
+    per100: { kcal: 197, protein_g: 11, fat_g: 15, carbs_g: 5 },
+  },
+  {
+    name: "Грецький салат",
+    per100: { kcal: 99, protein_g: 3, fat_g: 8, carbs_g: 5 },
+  },
+  {
+    name: "Капрезе",
+    per100: { kcal: 212, protein_g: 11, fat_g: 17, carbs_g: 3 },
+  },
+  {
+    name: "Салат з тунцем",
+    per100: { kcal: 135, protein_g: 14, fat_g: 7, carbs_g: 4 },
+  },
+  {
+    name: "Салат із буряка",
+    per100: { kcal: 72, protein_g: 1.5, fat_g: 4, carbs_g: 8 },
+  },
+];
