+++ apps/web/src/modules/nutrition/lib/foodDb/seeds/saucesAndSpices.ts
@@ -1,0 +1,68 @@
+import type { SeedFood } from "./types";
+
+export const SAUCES_AND_SPICES: SeedFood[] = [
+  {
+    name: "Майонез",
+    per100: { kcal: 680, protein_g: 1, fat_g: 75, carbs_g: 2 },
+  },
+  {
+    name: "Кетчуп",
+    per100: { kcal: 100, protein_g: 1.5, fat_g: 0.1, carbs_g: 24 },
+  },
+  {
+    name: "Гірчиця",
+    per100: { kcal: 66, protein_g: 4, fat_g: 3.6, carbs_g: 6 },
+  },
+  {
+    name: "Соєвий соус",
+    per100: { kcal: 60, protein_g: 5.6, fat_g: 0, carbs_g: 8 },
+  },
+  {
+    name: "Томатна паста",
+    per100: { kcal: 82, protein_g: 3.9, fat_g: 0.5, carbs_g: 18 },
+  },
+  {
+    name: "Хумус",
+    per100: { kcal: 177, protein_g: 8, fat_g: 10, carbs_g: 16 },
+  },
+  {
+    name: "Песто",
+    per100: { kcal: 431, protein_g: 5, fat_g: 44, carbs_g: 6 },
+  },
+  {
+    name: "Соус терияки",
+    per100: { kcal: 90, protein_g: 5.9, fat_g: 0, carbs_g: 15 },
+  },
+  {
+    name: "Соус барбекю",
+    per100: { kcal: 172, protein_g: 1, fat_g: 0.5, carbs_g: 41 },
+  },
+  {
+    name: "Соус солодкий чилі",
+    per100: { kcal: 229, protein_g: 0.8, fat_g: 0.3, carbs_g: 56 },
+  },
+  {
+    name: "Соус тартар",
+    per100: { kcal: 389, protein_g: 1.4, fat_g: 38, carbs_g: 9.6 },
+  },
+  {
+    name: "Тахіні",
+    per100: { kcal: 595, protein_g: 17, fat_g: 54, carbs_g: 21 },
+  },
+  {
+    name: "Аджика",
+    per100: { kcal: 59, protein_g: 1, fat_g: 3.7, carbs_g: 5.8 },
+  },
+  {
+    name: "Ткемалі",
+    per100: { kcal: 89, protein_g: 1, fat_g: 0.2, carbs_g: 20 },
+  },
+  {
+    name: "Маргарин",
+    per100: { kcal: 717, protein_g: 0.2, fat_g: 80, carbs_g: 0.9 },
+  },
+  {
+    name: "Топлене масло (гі)",
+    per100: { kcal: 900, protein_g: 0, fat_g: 100, carbs_g: 0 },
+  },
+];
