+++ apps/web/src/modules/nutrition/lib/foodDb/seeds/sportsNutrition.ts
@@ -1,0 +1,28 @@
+import type { SeedFood } from "./types";
+
+export const SPORTS_NUTRITION: SeedFood[] = [
+  {
+    name: "Сироватковий протеїн",
+    per100: { kcal: 370, protein_g: 80, fat_g: 4, carbs_g: 7 },
+  },
+  {
+    name: "Казеїн",
+    per100: { kcal: 360, protein_g: 78, fat_g: 3, carbs_g: 8 },
+  },
+  {
+    name: "Гейнер",
+    per100: { kcal: 380, protein_g: 25, fat_g: 4, carbs_g: 65 },
+  },
+  {
+    name: "BCAA",
+    per100: { kcal: 340, protein_g: 85, fat_g: 0, carbs_g: 0 },
+  },
+  {
+    name: "Креатин",
+    per100: { kcal: 0, protein_g: 0, fat_g: 0, carbs_g: 0 },
+  },
+  {
+    name: "Ізотонік",
+    per100: { kcal: 24, protein_g: 0, fat_g: 0, carbs_g: 6 },
+  },
+];
