+++ apps/web/src/modules/nutrition/lib/foodDb/seeds/sweets.ts
@@ -1,0 +1,93 @@
+import type { SeedFood } from "./types";
+
+export const SWEETS: SeedFood[] = [
+  {
+    name: "Цукор",
+    per100: { kcal: 387, protein_g: 0, fat_g: 0, carbs_g: 100 },
+  },
+  { name: "Мед", per100: { kcal: 304, protein_g: 0.3, fat_g: 0, carbs_g: 82 } },
+  {
+    name: "Варення",
+    per100: { kcal: 272, protein_g: 0.3, fat_g: 0.1, carbs_g: 70 },
+  },
+  {
+    name: "Чорний шоколад",
+    per100: { kcal: 546, protein_g: 4.9, fat_g: 31, carbs_g: 61 },
+  },
+  {
+    name: "Шоколад молочний",
+    per100: { kcal: 535, protein_g: 6.9, fat_g: 29, carbs_g: 60 },
+  },
+  {
+    name: "Зефір",
+    per100: { kcal: 304, protein_g: 0.8, fat_g: 0, carbs_g: 79 },
+  },
+  {
+    name: "Мармелад",
+    per100: { kcal: 322, protein_g: 0.4, fat_g: 0.1, carbs_g: 82 },
+  },
+  {
+    name: "Печиво вівсяне",
+    per100: { kcal: 437, protein_g: 6.5, fat_g: 14, carbs_g: 71 },
+  },
+  {
+    name: "Вафлі",
+    per100: { kcal: 514, protein_g: 8, fat_g: 28, carbs_g: 62 },
+  },
+  {
+    name: "Чіпси",
+    per100: { kcal: 520, protein_g: 6, fat_g: 29, carbs_g: 60 },
+  },
+  {
+    name: "Попкорн",
+    per100: { kcal: 375, protein_g: 11, fat_g: 4, carbs_g: 78 },
+  },
+  {
+    name: "Чизкейк",
+    per100: { kcal: 321, protein_g: 5.5, fat_g: 23, carbs_g: 26 },
+  },
+  {
+    name: "Тірамісу",
+    per100: { kcal: 240, protein_g: 5, fat_g: 14, carbs_g: 23 },
+  },
+  {
+    name: "Торт Наполеон",
+    per100: { kcal: 400, protein_g: 5, fat_g: 26, carbs_g: 38 },
+  },
+  {
+    name: "Торт Медовик",
+    per100: { kcal: 370, protein_g: 5, fat_g: 18, carbs_g: 48 },
+  },
+  {
+    name: "Торт Прага",
+    per100: { kcal: 387, protein_g: 4.5, fat_g: 21, carbs_g: 45 },
+  },
+  {
+    name: "Брауні",
+    per100: { kcal: 466, protein_g: 5, fat_g: 26, carbs_g: 55 },
+  },
+  {
+    name: "Молоко згущене",
+    per100: { kcal: 321, protein_g: 7.2, fat_g: 8.5, carbs_g: 56 },
+  },
+  {
+    name: "Халва",
+    per100: { kcal: 524, protein_g: 12, fat_g: 30, carbs_g: 54 },
+  },
+  {
+    name: "Курага в шоколаді",
+    per100: { kcal: 383, protein_g: 3.5, fat_g: 14, carbs_g: 64 },
+  },
+  {
+    name: "Карамель",
+    per100: { kcal: 385, protein_g: 0, fat_g: 0.1, carbs_g: 96 },
+  },
+  {
+    name: "Протеїновий батончик",
+    per100: { kcal: 380, protein_g: 30, fat_g: 14, carbs_g: 35 },
+  },
+  {
+    name: "Гранола батончик",
+    per100: { kcal: 400, protein_g: 6, fat_g: 14, carbs_g: 63 },
+  },
+];
