+++ apps/web/src/modules/nutrition/lib/foodDb/seeds/types.ts
@@ -1,0 +1,6 @@
+import type { Macros } from "../../macros";
+
+export interface SeedFood {
+  name: string;
+  per100: Macros;
+}
