+++ apps/web/src/modules/nutrition/lib/foodDb/seeds/ukrainianCuisine.ts
@@ -1,0 +1,76 @@
+import type { SeedFood } from "./types";
+
+export const UKRAINIAN_CUISINE: SeedFood[] = [
+  {
+    name: "Борщ зелений",
+    per100: { kcal: 49, protein_g: 2.5, fat_g: 2, carbs_g: 5 },
+  },
+  {
+    name: "Куряча локшина",
+    per100: { kcal: 45, protein_g: 3, fat_g: 1.5, carbs_g: 5 },
+  },
+  {
+    name: "Суп-харчо",
+    per100: { kcal: 86, protein_g: 6, fat_g: 4, carbs_g: 7 },
+  },
+  {
+    name: "Розсольник",
+    per100: { kcal: 49, protein_g: 2.5, fat_g: 2, carbs_g: 5.5 },
+  },
+  {
+    name: "Солянка м'ясна",
+    per100: { kcal: 105, protein_g: 6, fat_g: 7, carbs_g: 4 },
+  },
+  {
+    name: "Окрошка",
+    per100: { kcal: 53, protein_g: 2.7, fat_g: 2, carbs_g: 6 },
+  },
+  {
+    name: "Гороховий суп",
+    per100: { kcal: 66, protein_g: 4, fat_g: 2, carbs_g: 8 },
+  },
+  {
+    name: "Голубці",
+    per100: { kcal: 130, protein_g: 7, fat_g: 7, carbs_g: 10 },
+  },
+  {
+    name: "Деруни",
+    per100: { kcal: 170, protein_g: 3, fat_g: 10, carbs_g: 17 },
+  },
+  {
+    name: "Млинці",
+    per100: { kcal: 189, protein_g: 5, fat_g: 7, carbs_g: 27 },
+  },
+  {
+    name: "Налисники з сиром",
+    per100: { kcal: 219, protein_g: 9, fat_g: 10, carbs_g: 25 },
+  },
+  {
+    name: "Сирники",
+    per100: { kcal: 220, protein_g: 14, fat_g: 8, carbs_g: 23 },
+  },
+  {
+    name: "Галушки",
+    per100: { kcal: 165, protein_g: 5, fat_g: 3, carbs_g: 30 },
+  },
+  {
+    name: "Вареники з вишнею",
+    per100: { kcal: 165, protein_g: 4, fat_g: 1, carbs_g: 35 },
+  },
+  {
+    name: "Вареники з м'ясом",
+    per100: { kcal: 244, protein_g: 13, fat_g: 8, carbs_g: 30 },
+  },
+  {
+    name: "Котлета домашня",
+    per100: { kcal: 221, protein_g: 14, fat_g: 14, carbs_g: 10 },
+  },
+  {
+    name: "Кабачкова ікра",
+    per100: { kcal: 91, protein_g: 1.3, fat_g: 6.8, carbs_g: 6.4 },
+  },
+  {
+    name: "Баклажанна ікра",
+    per100: { kcal: 88, protein_g: 1.7, fat_g: 6.7, carbs_g: 6.1 },
+  },
+];
