+++ apps/web/src/modules/nutrition/lib/foodDb/seeds/vegetables.ts
@@ -1,0 +1,188 @@
+import type { SeedFood } from "./types";
+
+export const VEGETABLES: SeedFood[] = [
+  {
+    name: "Картопля",
+    per100: { kcal: 77, protein_g: 2, fat_g: 0.1, carbs_g: 17 },
+  },
+  {
+    name: "Батат",
+    per100: { kcal: 86, protein_g: 1.6, fat_g: 0.1, carbs_g: 20 },
+  },
+  {
+    name: "Томат",
+    per100: { kcal: 18, protein_g: 0.9, fat_g: 0.2, carbs_g: 3.9 },
+  },
+  {
+    name: "Огірок",
+    per100: { kcal: 15, protein_g: 0.7, fat_g: 0.1, carbs_g: 3.6 },
+  },
+  {
+    name: "Перець солодкий",
+    per100: { kcal: 31, protein_g: 1, fat_g: 0.3, carbs_g: 6 },
+  },
+  {
+    name: "Морква",
+    per100: { kcal: 41, protein_g: 0.9, fat_g: 0.2, carbs_g: 10 },
+  },
+  {
+    name: "Цибуля",
+    per100: { kcal: 40, protein_g: 1.1, fat_g: 0.1, carbs_g: 9.3 },
+  },
+  {
+    name: "Цибуля зелена",
+    per100: { kcal: 32, protein_g: 1.8, fat_g: 0.2, carbs_g: 7 },
+  },
+  {
+    name: "Часник",
+    per100: { kcal: 149, protein_g: 6.4, fat_g: 0.5, carbs_g: 33 },
+  },
+  {
+    name: "Капуста білокачанна",
+    per100: { kcal: 25, protein_g: 1.3, fat_g: 0.1, carbs_g: 5.8 },
+  },
+  {
+    name: "Капуста пекінська",
+    per100: { kcal: 16, protein_g: 1.2, fat_g: 0.2, carbs_g: 2.5 },
+  },
+  {
+    name: "Цвітна капуста",
+    per100: { kcal: 25, protein_g: 1.9, fat_g: 0.3, carbs_g: 5 },
+  },
+  {
+    name: "Броколі",
+    per100: { kcal: 34, protein_g: 2.8, fat_g: 0.4, carbs_g: 7 },
+  },
+  {
+    name: "Шпинат",
+    per100: { kcal: 23, protein_g: 2.9, fat_g: 0.4, carbs_g: 3.6 },
+  },
+  {
+    name: "Салат листовий",
+    per100: { kcal: 15, protein_g: 1.4, fat_g: 0.2, carbs_g: 2.9 },
+  },
+  {
+    name: "Руккола",
+    per100: { kcal: 25, protein_g: 2.6, fat_g: 0.7, carbs_g: 3.7 },
+  },
+  {
+    name: "Буряк",
+    per100: { kcal: 43, protein_g: 1.6, fat_g: 0.2, carbs_g: 10 },
+  },
+  {
+    name: "Кабачок",
+    per100: { kcal: 17, protein_g: 1.2, fat_g: 0.3, carbs_g: 3.1 },
+  },
+  {
+    name: "Баклажан",
+    per100: { kcal: 25, protein_g: 1, fat_g: 0.2, carbs_g: 5.7 },
+  },
+  {
+    name: "Гарбуз",
+    per100: { kcal: 26, protein_g: 1, fat_g: 0.1, carbs_g: 6.5 },
+  },
+  {
+    name: "Кукурудза",
+    per100: { kcal: 86, protein_g: 3.3, fat_g: 1.4, carbs_g: 19 },
+  },
+  {
+    name: "Гриби (шампіньони)",
+    per100: { kcal: 22, protein_g: 3.1, fat_g: 0.3, carbs_g: 3.3 },
+  },
+  {
+    name: "Редиска",
+    per100: { kcal: 16, protein_g: 0.7, fat_g: 0.1, carbs_g: 3.4 },
+  },
+  {
+    name: "Петрушка",
+    per100: { kcal: 36, protein_g: 3, fat_g: 0.8, carbs_g: 6.3 },
+  },
+  {
+    name: "Кріп",
+    per100: { kcal: 43, protein_g: 3.5, fat_g: 1.1, carbs_g: 7 },
+  },
+  {
+    name: "Квашена капуста",
+    per100: { kcal: 19, protein_g: 0.9, fat_g: 0.1, carbs_g: 4 },
+  },
+  {
+    name: "Авокадо",
+    per100: { kcal: 160, protein_g: 2, fat_g: 15, carbs_g: 9 },
+  },
+  {
+    name: "Спаржа",
+    per100: { kcal: 20, protein_g: 2.2, fat_g: 0.1, carbs_g: 3.9 },
+  },
+  {
+    name: "Імбир свіжий",
+    per100: { kcal: 80, protein_g: 1.8, fat_g: 0.8, carbs_g: 18 },
+  },
+  {
+    name: "Селера (черешковий)",
+    per100: { kcal: 16, protein_g: 0.7, fat_g: 0.2, carbs_g: 3 },
+  },
+  {
+    name: "Селера (корінь)",
+    per100: { kcal: 42, protein_g: 1.5, fat_g: 0.3, carbs_g: 9.2 },
+  },
+  {
+    name: "Ріпа",
+    per100: { kcal: 28, protein_g: 0.9, fat_g: 0.1, carbs_g: 6.4 },
+  },
+  {
+    name: "Редька",
+    per100: { kcal: 20, protein_g: 1.2, fat_g: 0.2, carbs_g: 3.4 },
+  },
+  {
+    name: "Перець чилі",
+    per100: { kcal: 40, protein_g: 1.9, fat_g: 0.4, carbs_g: 8.8 },
+  },
+  {
+    name: "Гриби білі",
+    per100: { kcal: 34, protein_g: 3.7, fat_g: 1.7, carbs_g: 1.1 },
+  },
+  {
+    name: "Гриби лисички",
+    per100: { kcal: 38, protein_g: 1.5, fat_g: 1.1, carbs_g: 6.9 },
+  },
+  {
+    name: "Гриби печериці мариновані",
+    per100: { kcal: 24, protein_g: 2.8, fat_g: 0.4, carbs_g: 0.1 },
+  },
+  {
+    name: "Помідор черрі",
+    per100: { kcal: 18, protein_g: 0.9, fat_g: 0.2, carbs_g: 3.9 },
+  },
+  {
+    name: "Томат в'ялений",
+    per100: { kcal: 258, protein_g: 14, fat_g: 3, carbs_g: 56 },
+  },
+  {
+    name: "Кукурудза консервована",
+    per100: { kcal: 58, protein_g: 2.2, fat_g: 0.6, carbs_g: 11 },
+  },
+  {
+    name: "Горошок консервований",
+    per100: { kcal: 53, protein_g: 3.1, fat_g: 0.2, carbs_g: 9.8 },
+  },
+  {
+    name: "Маслини чорні",
+    per100: { kcal: 115, protein_g: 0.8, fat_g: 11, carbs_g: 6.3 },
+  },
+  {
+    name: "Оливки зелені",
+    per100: { kcal: 145, protein_g: 1, fat_g: 15, carbs_g: 3.8 },
+  },
+  {
+    name: "Каперси",
+    per100: { kcal: 23, protein_g: 2.4, fat_g: 0.9, carbs_g: 5 },
+  },
+  {
+    name: "Корнішони",
+    per100: { kcal: 15, protein_g: 0.7, fat_g: 0.1, carbs_g: 2.4 },
+  },
+  {
+    name: "Огірок солоний",
+    per100: { kcal: 11, protein_g: 0.8, fat_g: 0.1, carbs_g: 1.7 },
+  },
+];
